Removed unused Utility folder and fixed project references to use Utilities.
