+++ BuildCostEstimator.DataAccess/BuildCostEstimator.DataAccess.csproj
@@ -14,7 +14,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\BuildCostEstimator.Models\BuildCostEstimator.Models.csproj" />
-    <ProjectReference Include="..\BuildCostEstimator.Utility\BuildCostEstimator.Utilities.csproj" />
+    <ProjectReference Include="..\BuildCostEstimator.Utilities\BuildCostEstimator.Utilities.csproj" />
   </ItemGroup>
 
 </Project>
