+++ BuildCostEstimator.PriceCheck/UniquePriceChecker.cs
@@ -13,7 +13,7 @@ using System.Threading.Tasks;
 using BuildCostEstimator.Models;
 using BuildCostEstimator.Models.PoeNinjaModels;
 using BuildCostEstimator.PriceCheck.Interfaces;
-using BuildCostEstimator.Utilities;
+using BuildCostEstimator.Utilities.Extensions;
 using Microsoft.Extensions.Caching.Memory;
