+++ BuildCostEstimator.Utilities/BuildCostEstimator.Utilities.csproj
@@ -4,4 +4,8 @@
     <TargetFramework>net5.0</TargetFramework>
   </PropertyGroup>
 
+  <ItemGroup>
+    <ProjectReference Include="..\BuildCostEstimator.Models\BuildCostEstimator.Models.csproj" />
+  </ItemGroup>
+
 </Project>
