+++ BuildCostEstimator.Utilities/PastebinDataService.cs
