+++ BuildCostEstimator.Utilities/StaticDetails.cs
@@ -1,12 +1,2732 @@
 ﻿using System;
 using System.Collections.Generic;
+using System.Dynamic;
 using System.Linq;
 using System.Text;
 using System.Threading.Tasks;
 
 namespace BuildCostEstimator.Utilities
 {
-    class StaticDetails
+    public static class StaticDetails
     {
+        
+        /*
+         Item Class: Amulets
+Rarity: Rare
+Dragon Torc
+Amber Amulet
+--------
+Requirements:
+Level: 9
+--------
+Item Level: 16
+--------
++28 to Strength (implicit)
+--------
+5% increased Cold Damage
+4% increased Armour
++19 to maximum Mana
++8% to Lightning Resistance
++2 Life gained for each Enemy hit by your Attacks
+*/
+
+
+        public static Dictionary<string, string> BaseTypeToItemClass = new()
+        {
+            {"Blue Pearl Amulet", "Amulet"},
+            {"Marble Amulet", "Amulet"},
+            {"Seaglass Amulet", "Amulet"},
+            {"Jet Amulet", "Amulet"},
+            {"Paua Amulet", "Amulet"},
+            {"Citrine Amulet", "Amulet"},
+            {"Ruby Amulet", "Amulet"},
+            {"Coral Amulet", "Amulet"},
+            {"Amber Amulet", "Amulet"},
+            {"Jade Amulet", "Amulet"},
+            {"Lapis Amulet", "Amulet"},
+            {"Gold Amulet", "Amulet"},
+            {"Onyx Amulet", "Amulet"},
+            {"Turquoise Amulet", "Amulet"},
+            {"Agate Amulet", "Amulet"},
+            {"Astrolabe Amulet", "Amulet"},
+            {"Simplex Amulet", "Amulet"},
+            {"Black Maw Talisman", "Amulet"},
+            {"Mandible Talisman", "Amulet"},
+            {"Chrysalis Talisman", "Amulet"},
+            {"Writhing Talisman", "Amulet"},
+            {"Bonespire Talisman", "Amulet"},
+            {"Ashscale Talisman", "Amulet"},
+            {"Lone Antler Talisman", "Amulet"},
+            {"Deep One Talisman", "Amulet"},
+            {"Breakrib Talisman", "Amulet"},
+            {"Deadhand Talisman", "Amulet"},
+            {"Undying Flesh Talisman", "Amulet"},
+            {"Rot Head Talisman", "Amulet"},
+            {"Hexclaw Talisman", "Amulet"},
+            {"Primal Skull Talisman", "Amulet"},
+            {"Wereclaw Talisman", "Amulet"},
+            {"Splitnewt Talisman", "Amulet"},
+            {"Clutching Talisman", "Amulet"},
+            {"Avian Twins Talisman", "Amulet"},
+            {"Fangjaw Talisman", "Amulet"},
+            {"Horned Talisman", "Amulet"},
+            {"Spinefuse Talisman", "Amulet"},
+            {"Three Rat Talisman", "Amulet"},
+            {"Monkey Twins Talisman", "Amulet"},
+            {"Longtooth Talisman", "Amulet"},
+            {"Rotfeather Talisman", "Amulet"},
+            {"Monkey Paw Talisman", "Amulet"},
+            {"Three Hands Talisman", "Amulet"},
+            {"Greatwolf Talisman", "Amulet"},
+            {"Golden Mantle", "Body Armour"},
+            {"Shabby Jerkin", "Body Armour"},
+            {"Glorious Leather", "Body Armour"},
+            {"Coronal Leather", "Body Armour"},
+            {"Cutthroat's Garb", "Body Armour"},
+            {"Sharkskin Tunic", "Body Armour"},
+            {"Destiny Leather", "Body Armour"},
+            {"Exquisite Leather", "Body Armour"},
+            {"Zodiac Leather", "Body Armour"},
+            {"Assassin's Garb", "Body Armour"},
+            {"Strapped Leather", "Body Armour"},
+            {"Buckskin Tunic", "Body Armour"},
+            {"Wild Leather", "Body Armour"},
+            {"Full Leather", "Body Armour"},
+            {"Sun Leather", "Body Armour"},
+            {"Thief's Garb", "Body Armour"},
+            {"Eelskin Tunic", "Body Armour"},
+            {"Frontier Leather", "Body Armour"},
+            {"Padded Vest", "Body Armour"},
+            {"Crimson Raiment", "Body Armour"},
+            {"Lacquered Garb", "Body Armour"},
+            {"Crypt Armour", "Body Armour"},
+            {"Sentinel Jacket", "Body Armour"},
+            {"Varnished Coat", "Body Armour"},
+            {"Blood Raiment", "Body Armour"},
+            {"Sadist Garb", "Body Armour"},
+            {"Carnal Armour", "Body Armour"},
+            {"Oiled Vest", "Body Armour"},
+            {"Padded Jacket", "Body Armour"},
+            {"Oiled Coat", "Body Armour"},
+            {"Scarlet Raiment", "Body Armour"},
+            {"Waxed Garb", "Body Armour"},
+            {"Bone Armour", "Body Armour"},
+            {"Quilted Jacket", "Body Armour"},
+            {"Sleek Coat", "Body Armour"},
+            {"Simple Robe", "Body Armour"},
+            {"Conjurer's Vestment", "Body Armour"},
+            {"Spidersilk Robe", "Body Armour"},
+            {"Destroyer Regalia", "Body Armour"},
+            {"Savant's Robe", "Body Armour"},
+            {"Necromancer Silks", "Body Armour"},
+            {"Occultist's Vestment", "Body Armour"},
+            {"Widowsilk Robe", "Body Armour"},
+            {"Vaal Regalia", "Body Armour"},
+            {"Silken Vest", "Body Armour"},
+            {"Scholar's Robe", "Body Armour"},
+            {"Silken Garb", "Body Armour"},
+            {"Mage's Vestment", "Body Armour"},
+            {"Silk Robe", "Body Armour"},
+            {"Cabalist Regalia", "Body Armour"},
+            {"Sage's Robe", "Body Armour"},
+            {"Silken Wrap", "Body Armour"},
+            {"Plate Vest", "Body Armour"},
+            {"Sun Plate", "Body Armour"},
+            {"Colosseum Plate", "Body Armour"},
+            {"Majestic Plate", "Body Armour"},
+            {"Golden Plate", "Body Armour"},
+            {"Crusader Plate", "Body Armour"},
+            {"Astral Plate", "Body Armour"},
+            {"Gladiator Plate", "Body Armour"},
+            {"Glorious Plate", "Body Armour"},
+            {"Chestplate", "Body Armour"},
+            {"Copper Plate", "Body Armour"},
+            {"War Plate", "Body Armour"},
+            {"Full Plate", "Body Armour"},
+            {"Arena Plate", "Body Armour"},
+            {"Lordly Plate", "Body Armour"},
+            {"Bronze Plate", "Body Armour"},
+            {"Battle Plate", "Body Armour"},
+            {"Scale Vest", "Body Armour"},
+            {"Full Wyrmscale", "Body Armour"},
+            {"Commander's Brigandine", "Body Armour"},
+            {"Battle Lamellar", "Body Armour"},
+            {"Dragonscale Doublet", "Body Armour"},
+            {"Desert Brigandine", "Body Armour"},
+            {"Full Dragonscale", "Body Armour"},
+            {"General's Brigandine", "Body Armour"},
+            {"Triumphant Lamellar", "Body Armour"},
+            {"Light Brigandine", "Body Armour"},
+            {"Scale Doublet", "Body Armour"},
+            {"Infantry Brigandine", "Body Armour"},
+            {"Full Scale Armour", "Body Armour"},
+            {"Soldier's Brigandine", "Body Armour"},
+            {"Field Lamellar", "Body Armour"},
+            {"Wyrmscale Doublet", "Body Armour"},
+            {"Hussar Brigandine", "Body Armour"},
+            {"Sacrificial Garb", "Body Armour"},
+            {"Grasping Mail", "Body Armour"},
+            {"Chainmail Vest", "Body Armour"},
+            {"Ornate Ringmail", "Body Armour"},
+            {"Chain Hauberk", "Body Armour"},
+            {"Devout Chainmail", "Body Armour"},
+            {"Loricated Ringmail", "Body Armour"},
+            {"Conquest Chainmail", "Body Armour"},
+            {"Elegant Ringmail", "Body Armour"},
+            {"Saint's Hauberk", "Body Armour"},
+            {"Saintly Chainmail", "Body Armour"},
+            {"Chainmail Tunic", "Body Armour"},
+            {"Ringmail Coat", "Body Armour"},
+            {"Chainmail Doublet", "Body Armour"},
+            {"Full Ringmail", "Body Armour"},
+            {"Full Chainmail", "Body Armour"},
+            {"Holy Chainmail", "Body Armour"},
+            {"Latticed Ringmail", "Body Armour"},
+            {"Crusader Chainmail", "Body Armour"},
+            {"Kaom's Plate", "Body Armour"},
+            {"Two-Toned Boots", "Boots"},
+            {"Blessed Boots", "Boots"},
+            {"Golden Caligae", "Boots"},
+            {"Avian Slippers", "Boots"},
+            {"Rawhide Boots", "Boots"},
+            {"Goathide Boots", "Boots"},
+            {"Deerskin Boots", "Boots"},
+            {"Nubuck Boots", "Boots"},
+            {"Eelskin Boots", "Boots"},
+            {"Sharkskin Boots", "Boots"},
+            {"Shagreen Boots", "Boots"},
+            {"Stealth Boots", "Boots"},
+            {"Slink Boots", "Boots"},
+            {"Wrapped Boots", "Boots"},
+            {"Strapped Boots", "Boots"},
+            {"Clasped Boots", "Boots"},
+            {"Shackled Boots", "Boots"},
+            {"Trapper Boots", "Boots"},
+            {"Ambush Boots", "Boots"},
+            {"Carnal Boots", "Boots"},
+            {"Assassin's Boots", "Boots"},
+            {"Murder Boots", "Boots"},
+            {"Cloudwhisper Boots", "Boots"},
+            {"Windbreak Boots", "Boots"},
+            {"Stormrider Boots", "Boots"},
+            {"Wool Shoes", "Boots"},
+            {"Velvet Slippers", "Boots"},
+            {"Silk Slippers", "Boots"},
+            {"Scholar Boots", "Boots"},
+            {"Satin Slippers", "Boots"},
+            {"Samite Slippers", "Boots"},
+            {"Conjurer Boots", "Boots"},
+            {"Arcanist Slippers", "Boots"},
+            {"Sorcerer Boots", "Boots"},
+            {"Duskwalk Slippers", "Boots"},
+            {"Nightwind Slippers", "Boots"},
+            {"Dreamquest Slippers", "Boots"},
+            {"Iron Greaves", "Boots"},
+            {"Steel Greaves", "Boots"},
+            {"Plated Greaves", "Boots"},
+            {"Reinforced Greaves", "Boots"},
+            {"Antique Greaves", "Boots"},
+            {"Ancient Greaves", "Boots"},
+            {"Goliath Greaves", "Boots"},
+            {"Vaal Greaves", "Boots"},
+            {"Titan Greaves", "Boots"},
+            {"Leatherscale Boots", "Boots"},
+            {"Ironscale Boots", "Boots"},
+            {"Bronzescale Boots", "Boots"},
+            {"Steelscale Boots", "Boots"},
+            {"Serpentscale Boots", "Boots"},
+            {"Wyrmscale Boots", "Boots"},
+            {"Hydrascale Boots", "Boots"},
+            {"Dragonscale Boots", "Boots"},
+            {"Chain Boots", "Boots"},
+            {"Ringmail Boots", "Boots"},
+            {"Mesh Boots", "Boots"},
+            {"Riveted Boots", "Boots"},
+            {"Zealot Boots", "Boots"},
+            {"Soldier Boots", "Boots"},
+            {"Legion Boots", "Boots"},
+            {"Crusader Boots", "Boots"},
+            {"Basemetal Treads", "Boots"},
+            {"Darksteel Treads", "Boots"},
+            {"Brimstone Treads", "Boots"},
+            {"Kaom's Greaves", "Boots"},
+            {"Gripped Gloves", "Gloves"},
+            {"Fingerless Silk Gloves", "Gloves"},
+            {"Spiked Gloves", "Gloves"},
+            {"Apothecary's Gloves", "Gloves"},
+            {"Golden Bracers", "Gloves"},
+            {"Rawhide Gloves", "Gloves"},
+            {"Goathide Gloves", "Gloves"},
+            {"Deerskin Gloves", "Gloves"},
+            {"Nubuck Gloves", "Gloves"},
+            {"Eelskin Gloves", "Gloves"},
+            {"Sharkskin Gloves", "Gloves"},
+            {"Shagreen Gloves", "Gloves"},
+            {"Stealth Gloves", "Gloves"},
+            {"Slink Gloves", "Gloves"},
+            {"Wrapped Mitts", "Gloves"},
+            {"Strapped Mitts", "Gloves"},
+            {"Clasped Mitts", "Gloves"},
+            {"Trapper Mitts", "Gloves"},
+            {"Ambush Mitts", "Gloves"},
+            {"Carnal Mitts", "Gloves"},
+            {"Assassin's Mitts", "Gloves"},
+            {"Murder Mitts", "Gloves"},
+            {"Gauche Gloves", "Gloves"},
+            {"Southswing Gloves", "Gloves"},
+            {"Sinistral Gloves", "Gloves"},
+            {"Wool Gloves", "Gloves"},
+            {"Velvet Gloves", "Gloves"},
+            {"Silk Gloves", "Gloves"},
+            {"Embroidered Gloves", "Gloves"},
+            {"Satin Gloves", "Gloves"},
+            {"Samite Gloves", "Gloves"},
+            {"Conjurer Gloves", "Gloves"},
+            {"Arcanist Gloves", "Gloves"},
+            {"Sorcerer Gloves", "Gloves"},
+            {"Leyline Gloves", "Gloves"},
+            {"Aetherwind Gloves", "Gloves"},
+            {"Nexus Gloves", "Gloves"},
+            {"Iron Gauntlets", "Gloves"},
+            {"Plated Gauntlets", "Gloves"},
+            {"Bronze Gauntlets", "Gloves"},
+            {"Steel Gauntlets", "Gloves"},
+            {"Antique Gauntlets", "Gloves"},
+            {"Ancient Gauntlets", "Gloves"},
+            {"Goliath Gauntlets", "Gloves"},
+            {"Vaal Gauntlets", "Gloves"},
+            {"Titan Gauntlets", "Gloves"},
+            {"Fishscale Gauntlets", "Gloves"},
+            {"Ironscale Gauntlets", "Gloves"},
+            {"Bronzescale Gauntlets", "Gloves"},
+            {"Steelscale Gauntlets", "Gloves"},
+            {"Serpentscale Gauntlets", "Gloves"},
+            {"Wyrmscale Gauntlets", "Gloves"},
+            {"Hydrascale Gauntlets", "Gloves"},
+            {"Dragonscale Gauntlets", "Gloves"},
+            {"Chain Gloves", "Gloves"},
+            {"Ringmail Gloves", "Gloves"},
+            {"Mesh Gloves", "Gloves"},
+            {"Riveted Gloves", "Gloves"},
+            {"Zealot Gloves", "Gloves"},
+            {"Soldier Gloves", "Gloves"},
+            {"Legion Gloves", "Gloves"},
+            {"Crusader Gloves", "Gloves"},
+            {"Taxing Gauntlets", "Gloves"},
+            {"Gruelling Gauntlets", "Gloves"},
+            {"Debilitation Gauntlets", "Gloves"},
+            {"Bone Helmet", "Helmet"},
+            {"Golden Visage", "Helmet"},
+            {"Leather Cap", "Helmet"},
+            {"Lion Pelt", "Helmet"},
+            {"Tricorne", "Helmet"},
+            {"Leather Hood", "Helmet"},
+            {"Wolf Pelt", "Helmet"},
+            {"Hunter Hood", "Helmet"},
+            {"Noble Tricorne", "Helmet"},
+            {"Ursine Pelt", "Helmet"},
+            {"Silken Hood", "Helmet"},
+            {"Sinner Tricorne", "Helmet"},
+            {"Scare Mask", "Helmet"},
+            {"Vaal Mask", "Helmet"},
+            {"Deicide Mask", "Helmet"},
+            {"Plague Mask", "Helmet"},
+            {"Iron Mask", "Helmet"},
+            {"Festival Mask", "Helmet"},
+            {"Golden Mask", "Helmet"},
+            {"Raven Mask", "Helmet"},
+            {"Callous Mask", "Helmet"},
+            {"Regicide Mask", "Helmet"},
+            {"Harlequin Mask", "Helmet"},
+            {"Gale Crown", "Helmet"},
+            {"Winter Crown", "Helmet"},
+            {"Blizzard Crown", "Helmet"},
+            {"Vine Circlet", "Helmet"},
+            {"Mind Cage", "Helmet"},
+            {"Hubris Circlet", "Helmet"},
+            {"Iron Circlet", "Helmet"},
+            {"Torture Cage", "Helmet"},
+            {"Tribal Circlet", "Helmet"},
+            {"Bone Circlet", "Helmet"},
+            {"Lunaris Circlet", "Helmet"},
+            {"Steel Circlet", "Helmet"},
+            {"Necromancer Circlet", "Helmet"},
+            {"Solaris Circlet", "Helmet"},
+            {"Iron Hat", "Helmet"},
+            {"Royal Burgonet", "Helmet"},
+            {"Eternal Burgonet", "Helmet"},
+            {"Cone Helmet", "Helmet"},
+            {"Barbute Helmet", "Helmet"},
+            {"Close Helmet", "Helmet"},
+            {"Gladiator Helmet", "Helmet"},
+            {"Reaver Helmet", "Helmet"},
+            {"Siege Helmet", "Helmet"},
+            {"Samnite Helmet", "Helmet"},
+            {"Ezomyte Burgonet", "Helmet"},
+            {"Battered Helm", "Helmet"},
+            {"Nightmare Bascinet", "Helmet"},
+            {"Sallet", "Helmet"},
+            {"Visored Sallet", "Helmet"},
+            {"Gilded Sallet", "Helmet"},
+            {"Secutor Helm", "Helmet"},
+            {"Fencer Helm", "Helmet"},
+            {"Lacquered Helmet", "Helmet"},
+            {"Fluted Bascinet", "Helmet"},
+            {"Pig-Faced Bascinet", "Helmet"},
+            {"Sorrow Mask", "Helmet"},
+            {"Atonement Mask", "Helmet"},
+            {"Penitent Mask", "Helmet"},
+            {"Rusted Coif", "Helmet"},
+            {"Praetor Crown", "Helmet"},
+            {"Soldier Helmet", "Helmet"},
+            {"Great Helmet", "Helmet"},
+            {"Crusader Helmet", "Helmet"},
+            {"Aventail Helmet", "Helmet"},
+            {"Zealot Helmet", "Helmet"},
+            {"Great Crown", "Helmet"},
+            {"Magistrate Crown", "Helmet"},
+            {"Prophet Crown", "Helmet"},
+            {"Imp Crown", "Helmet"},
+            {"Demon Crown", "Helmet"},
+            {"Archdemon Crown", "Helmet"},
+            {"Golden Wreath", "Helmet"},
+            {"Golden Flame", "Shield"},
+            {"Goathide Buckler", "Shield"},
+            {"Battle Buckler", "Shield"},
+            {"Golden Buckler", "Shield"},
+            {"Ironwood Buckler", "Shield"},
+            {"Lacquered Buckler", "Shield"},
+            {"Vaal Buckler", "Shield"},
+            {"Crusader Buckler", "Shield"},
+            {"Imperial Buckler", "Shield"},
+            {"Pine Buckler", "Shield"},
+            {"Painted Buckler", "Shield"},
+            {"Hammered Buckler", "Shield"},
+            {"War Buckler", "Shield"},
+            {"Gilded Buckler", "Shield"},
+            {"Oak Buckler", "Shield"},
+            {"Enameled Buckler", "Shield"},
+            {"Corrugated Buckler", "Shield"},
+            {"Endothermic Buckler", "Shield"},
+            {"Polar Buckler", "Shield"},
+            {"Cold-attuned Buckler", "Shield"},
+            {"Spiked Bundle", "Shield"},
+            {"Alder Spiked Shield", "Shield"},
+            {"Ezomyte Spiked Shield", "Shield"},
+            {"Mirrored Spiked Shield", "Shield"},
+            {"Supreme Spiked Shield", "Shield"},
+            {"Driftwood Spiked Shield", "Shield"},
+            {"Alloyed Spiked Shield", "Shield"},
+            {"Burnished Spiked Shield", "Shield"},
+            {"Ornate Spiked Shield", "Shield"},
+            {"Redwood Spiked Shield", "Shield"},
+            {"Compound Spiked Shield", "Shield"},
+            {"Polished Spiked Shield", "Shield"},
+            {"Sovereign Spiked Shield", "Shield"},
+            {"Twig Spirit Shield", "Shield"},
+            {"Chiming Spirit Shield", "Shield"},
+            {"Thorium Spirit Shield", "Shield"},
+            {"Lacewood Spirit Shield", "Shield"},
+            {"Fossilised Spirit Shield", "Shield"},
+            {"Vaal Spirit Shield", "Shield"},
+            {"Harmonic Spirit Shield", "Shield"},
+            {"Titanium Spirit Shield", "Shield"},
+            {"Yew Spirit Shield", "Shield"},
+            {"Bone Spirit Shield", "Shield"},
+            {"Tarnished Spirit Shield", "Shield"},
+            {"Jingling Spirit Shield", "Shield"},
+            {"Brass Spirit Shield", "Shield"},
+            {"Walnut Spirit Shield", "Shield"},
+            {"Ivory Spirit Shield", "Shield"},
+            {"Ancient Spirit Shield", "Shield"},
+            {"Exhausting Spirit Shield", "Shield"},
+            {"Subsuming Spirit Shield", "Shield"},
+            {"Transfer-attuned Spirit Shield", "Shield"},
+            {"Splintered Tower Shield", "Shield"},
+            {"Bronze Tower Shield", "Shield"},
+            {"Girded Tower Shield", "Shield"},
+            {"Crested Tower Shield", "Shield"},
+            {"Shagreen Tower Shield", "Shield"},
+            {"Ebony Tower Shield", "Shield"},
+            {"Ezomyte Tower Shield", "Shield"},
+            {"Colossal Tower Shield", "Shield"},
+            {"Pinnacle Tower Shield", "Shield"},
+            {"Corroded Tower Shield", "Shield"},
+            {"Rawhide Tower Shield", "Shield"},
+            {"Cedar Tower Shield", "Shield"},
+            {"Copper Tower Shield", "Shield"},
+            {"Reinforced Tower Shield", "Shield"},
+            {"Painted Tower Shield", "Shield"},
+            {"Buckskin Tower Shield", "Shield"},
+            {"Mahogany Tower Shield", "Shield"},
+            {"Rotted Round Shield", "Shield"},
+            {"Teak Round Shield", "Shield"},
+            {"Spiny Round Shield", "Shield"},
+            {"Cardinal Round Shield", "Shield"},
+            {"Elegant Round Shield", "Shield"},
+            {"Fir Round Shield", "Shield"},
+            {"Studded Round Shield", "Shield"},
+            {"Scarlet Round Shield", "Shield"},
+            {"Splendid Round Shield", "Shield"},
+            {"Maple Round Shield", "Shield"},
+            {"Spiked Round Shield", "Shield"},
+            {"Crimson Round Shield", "Shield"},
+            {"Baroque Round Shield", "Shield"},
+            {"Exothermic Tower Shield", "Shield"},
+            {"Magmatic Tower Shield", "Shield"},
+            {"Heat-attuned Tower Shield", "Shield"},
+            {"Plank Kite Shield", "Shield"},
+            {"Branded Kite Shield", "Shield"},
+            {"Champion Kite Shield", "Shield"},
+            {"Mosaic Kite Shield", "Shield"},
+            {"Archon Kite Shield", "Shield"},
+            {"Linden Kite Shield", "Shield"},
+            {"Reinforced Kite Shield", "Shield"},
+            {"Layered Kite Shield", "Shield"},
+            {"Ceremonial Kite Shield", "Shield"},
+            {"Etched Kite Shield", "Shield"},
+            {"Steel Kite Shield", "Shield"},
+            {"Laminated Kite Shield", "Shield"},
+            {"Angelic Kite Shield", "Shield"},
+            {"Crusader's Exalted Orb", "StackableCurrency"},
+            {"Hunter's Exalted Orb", "StackableCurrency"},
+            {"Redeemer's Exalted Orb", "StackableCurrency"},
+            {"Warlord's Exalted Orb", "StackableCurrency"},
+            {"Awakener's Orb", "StackableCurrency"},
+            {"Crimson Watchstone", "AtlasRegionUpgradeItem"},
+            {"Cobalt Watchstone", "AtlasRegionUpgradeItem"},
+            {"Viridian Watchstone", "AtlasRegionUpgradeItem"},
+            {"Golden Watchstone", "AtlasRegionUpgradeItem"},
+            {"Ivory Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium Haewark Hamlet Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium Tirn's End Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium Lex Proxima Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium Lex Ejoris Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium New Vastir Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium Glennach Cairns Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium Valdo's Rest Watchstone", "AtlasRegionUpgradeItem"},
+            {"Chromium Lira Arthain Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum Haewark Hamlet Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum Tirn's End Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum Lex Proxima Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum Lex Ejoris Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum New Vastir Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum Glennach Cairns Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum Valdo's Rest Watchstone", "AtlasRegionUpgradeItem"},
+            {"Platinum Lira Arthain Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium Haewark Hamlet Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium Tirn's End Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium Lex Proxima Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium Lex Ejoris Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium New Vastir Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium Glennach Cairns Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium Valdo's Rest Watchstone", "AtlasRegionUpgradeItem"},
+            {"Titanium Lira Arthain Watchstone", "AtlasRegionUpgradeItem"},
+            {"Rustic Sash", "Belt"},
+            {"Chain Belt", "Belt"},
+            {"Leather Belt", "Belt"},
+            {"Heavy Belt", "Belt"},
+            {"Cloth Belt", "Belt"},
+            {"Studded Belt", "Belt"},
+            {"Stygian Vise", "Belt"},
+            {"Vanguard Belt", "Belt"},
+            {"Crystal Belt", "Belt"},
+            {"Golden Obi", "Belt"},
+            {"Micro-Distillery Belt", "Belt"},
+            {"Mechalarm Belt", "Belt"},
+            {"Mystery Leaguestone", "Currency"},
+            {"Simple Rope Net", "StackableCurrency"},
+            {"Thaumaturgical Net", "StackableCurrency"},
+            {"Necromancy Net", "StackableCurrency"},
+            {"Reinforced Rope Net", "StackableCurrency"},
+            {"Strong Rope Net", "StackableCurrency"},
+            {"Simple Iron Net", "StackableCurrency"},
+            {"Reinforced Iron Net", "StackableCurrency"},
+            {"Strong Iron Net", "StackableCurrency"},
+            {"Simple Steel Net", "StackableCurrency"},
+            {"Reinforced Steel Net", "StackableCurrency"},
+            {"Strong Steel Net", "StackableCurrency"},
+            {"Simple Sextant", "StackableCurrency"},
+            {"Awakened Sextant", "StackableCurrency"},
+            {"Elevated Sextant", "StackableCurrency"},
+            {"Prime Sextant", "StackableCurrency"},
+            {"Facetor's Lens", "StackableCurrency"},
+            {"Orb of Augmentation", "StackableCurrency"},
+            {"Exalted Orb", "StackableCurrency"},
+            {"Exalted Shard", "StackableCurrency"},
+            {"Abyssal Delirium Orb", "StackableCurrency"},
+            {"Armoursmith's Delirium Orb", "StackableCurrency"},
+            {"Blighted Delirium Orb", "StackableCurrency"},
+            {"Obscured Delirium Orb", "StackableCurrency"},
+            {"Fine Delirium Orb", "StackableCurrency"},
+            {"Diviner's Delirium Orb", "StackableCurrency"},
+            {"Whispering Delirium Orb", "StackableCurrency"},
+            {"Fossilised Delirium Orb", "StackableCurrency"},
+            {"Fragmented Delirium Orb", "StackableCurrency"},
+            {"Thaumaturge's Delirium Orb", "StackableCurrency"},
+            {"Delirium Orb", "StackableCurrency"},
+            {"Foreboding Delirium Orb", "StackableCurrency"},
+            {"Timeless Delirium Orb", "StackableCurrency"},
+            {"Imperial Delirium Orb", "StackableCurrency"},
+            {"Cartographer's Delirium Orb", "StackableCurrency"},
+            {"Amorphous Delirium Orb", "StackableCurrency"},
+            {"Decadent Delirium Orb", "StackableCurrency"},
+            {"Portentous Delirium Orb", "StackableCurrency"},
+            {"Skittering Delirium Orb", "StackableCurrency"},
+            {"Primal Delirium Orb", "StackableCurrency"},
+            {"Jeweller's Delirium Orb", "StackableCurrency"},
+            {"Singular Delirium Orb", "StackableCurrency"},
+            {"Blacksmith's Delirium Orb", "StackableCurrency"},
+            {"Simulacrum Splinter", "StackableCurrency"},
+            {"Armourer's Scrap", "StackableCurrency"},
+            {"Orb of Unmaking", "StackableCurrency"},
+            {"Splinter of Chayula", "StackableCurrency"},
+            {"Splinter of Tul", "StackableCurrency"},
+            {"Splinter of Xoph", "StackableCurrency"},
+            {"Splinter of Esh", "StackableCurrency"},
+            {"Splinter of Uul-Netol", "StackableCurrency"},
+            {"Blessing of Chayula", "StackableCurrency"},
+            {"Blessing of Tul", "StackableCurrency"},
+            {"Blessing of Xoph", "StackableCurrency"},
+            {"Blessing of Esh", "StackableCurrency"},
+            {"Blessing of Uul-Netol", "StackableCurrency"},
+            {"Orb of Scouring", "StackableCurrency"},
+            {"Vaal Orb", "StackableCurrency"},
+            {"Remnant of Corruption", "StackableCurrency"},
+            {"Hollow Fossil", "StackableCurrency"},
+            {"Serrated Fossil", "StackableCurrency"},
+            {"Corroded Fossil", "StackableCurrency"},
+            {"Aetheric Fossil", "StackableCurrency"},
+            {"Aberrant Fossil", "StackableCurrency"},
+            {"Frigid Fossil", "StackableCurrency"},
+            {"Glyphic Fossil", "StackableCurrency"},
+            {"Dense Fossil", "StackableCurrency"},
+            {"Prismatic Fossil", "StackableCurrency"},
+            {"Enchanted Fossil", "StackableCurrency"},
+            {"Scorched Fossil", "StackableCurrency"},
+            {"Faceted Fossil", "StackableCurrency"},
+            {"Pristine Fossil", "StackableCurrency"},
+            {"Metallic Fossil", "StackableCurrency"},
+            {"Sanctified Fossil", "StackableCurrency"},
+            {"Lucent Fossil", "StackableCurrency"},
+            {"Bound Fossil", "StackableCurrency"},
+            {"Fractured Fossil", "StackableCurrency"},
+            {"Jagged Fossil", "StackableCurrency"},
+            {"Perfect Fossil", "StackableCurrency"},
+            {"Tangled Fossil", "StackableCurrency"},
+            {"Gilded Fossil", "StackableCurrency"},
+            {"Encrusted Fossil", "StackableCurrency"},
+            {"Shuddering Fossil", "StackableCurrency"},
+            {"Bloodstained Fossil", "StackableCurrency"},
+            {"Mirror of Kalandra", "StackableCurrency"},
+            {"Mirror Shard", "StackableCurrency"},
+            {"Muttering Essence of Anger", "StackableCurrency"},
+            {"Weeping Essence of Anger", "StackableCurrency"},
+            {"Wailing Essence of Anger", "StackableCurrency"},
+            {"Screaming Essence of Anger", "StackableCurrency"},
+            {"Shrieking Essence of Anger", "StackableCurrency"},
+            {"Deafening Essence of Anger", "StackableCurrency"},
+            {"Wailing Essence of Anguish", "StackableCurrency"},
+            {"Screaming Essence of Anguish", "StackableCurrency"},
+            {"Shrieking Essence of Anguish", "StackableCurrency"},
+            {"Deafening Essence of Anguish", "StackableCurrency"},
+            {"Whispering Essence of Contempt", "StackableCurrency"},
+            {"Muttering Essence of Contempt", "StackableCurrency"},
+            {"Weeping Essence of Contempt", "StackableCurrency"},
+            {"Wailing Essence of Contempt", "StackableCurrency"},
+            {"Screaming Essence of Contempt", "StackableCurrency"},
+            {"Shrieking Essence of Contempt", "StackableCurrency"},
+            {"Deafening Essence of Contempt", "StackableCurrency"},
+            {"Essence of Delirium", "StackableCurrency"},
+            {"Weeping Essence of Doubt", "StackableCurrency"},
+            {"Wailing Essence of Doubt", "StackableCurrency"},
+            {"Screaming Essence of Doubt", "StackableCurrency"},
+            {"Shrieking Essence of Doubt", "StackableCurrency"},
+            {"Deafening Essence of Doubt", "StackableCurrency"},
+            {"Screaming Essence of Dread", "StackableCurrency"},
+            {"Shrieking Essence of Dread", "StackableCurrency"},
+            {"Deafening Essence of Dread", "StackableCurrency"},
+            {"Screaming Essence of Envy", "StackableCurrency"},
+            {"Shrieking Essence of Envy", "StackableCurrency"},
+            {"Deafening Essence of Envy", "StackableCurrency"},
+            {"Muttering Essence of Fear", "StackableCurrency"},
+            {"Weeping Essence of Fear", "StackableCurrency"},
+            {"Wailing Essence of Fear", "StackableCurrency"},
+            {"Screaming Essence of Fear", "StackableCurrency"},
+            {"Shrieking Essence of Fear", "StackableCurrency"},
+            {"Deafening Essence of Fear", "StackableCurrency"},
+            {"Whispering Essence of Greed", "StackableCurrency"},
+            {"Muttering Essence of Greed", "StackableCurrency"},
+            {"Weeping Essence of Greed", "StackableCurrency"},
+            {"Wailing Essence of Greed", "StackableCurrency"},
+            {"Screaming Essence of Greed", "StackableCurrency"},
+            {"Shrieking Essence of Greed", "StackableCurrency"},
+            {"Deafening Essence of Greed", "StackableCurrency"},
+            {"Whispering Essence of Hatred", "StackableCurrency"},
+            {"Muttering Essence of Hatred", "StackableCurrency"},
+            {"Weeping Essence of Hatred", "StackableCurrency"},
+            {"Wailing Essence of Hatred", "StackableCurrency"},
+            {"Screaming Essence of Hatred", "StackableCurrency"},
+            {"Shrieking Essence of Hatred", "StackableCurrency"},
+            {"Deafening Essence of Hatred", "StackableCurrency"},
+            {"Essence of Horror", "StackableCurrency"},
+            {"Essence of Hysteria", "StackableCurrency"},
+            {"Essence of Insanity", "StackableCurrency"},
+            {"Wailing Essence of Loathing", "StackableCurrency"},
+            {"Screaming Essence of Loathing", "StackableCurrency"},
+            {"Shrieking Essence of Loathing", "StackableCurrency"},
+            {"Deafening Essence of Loathing", "StackableCurrency"},
+            {"Screaming Essence of Misery", "StackableCurrency"},
+            {"Shrieking Essence of Misery", "StackableCurrency"},
+            {"Deafening Essence of Misery", "StackableCurrency"},
+            {"Weeping Essence of Rage", "StackableCurrency"},
+            {"Wailing Essence of Rage", "StackableCurrency"},
+            {"Screaming Essence of Rage", "StackableCurrency"},
+            {"Shrieking Essence of Rage", "StackableCurrency"},
+            {"Deafening Essence of Rage", "StackableCurrency"},
+            {"Screaming Essence of Scorn", "StackableCurrency"},
+            {"Shrieking Essence of Scorn", "StackableCurrency"},
+            {"Deafening Essence of Scorn", "StackableCurrency"},
+            {"Muttering Essence of Sorrow", "StackableCurrency"},
+            {"Weeping Essence of Sorrow", "StackableCurrency"},
+            {"Wailing Essence of Sorrow", "StackableCurrency"},
+            {"Screaming Essence of Sorrow", "StackableCurrency"},
+            {"Shrieking Essence of Sorrow", "StackableCurrency"},
+            {"Deafening Essence of Sorrow", "StackableCurrency"},
+            {"Wailing Essence of Spite", "StackableCurrency"},
+            {"Screaming Essence of Spite", "StackableCurrency"},
+            {"Shrieking Essence of Spite", "StackableCurrency"},
+            {"Deafening Essence of Spite", "StackableCurrency"},
+            {"Weeping Essence of Suffering", "StackableCurrency"},
+            {"Wailing Essence of Suffering", "StackableCurrency"},
+            {"Screaming Essence of Suffering", "StackableCurrency"},
+            {"Shrieking Essence of Suffering", "StackableCurrency"},
+            {"Deafening Essence of Suffering", "StackableCurrency"},
+            {"Muttering Essence of Torment", "StackableCurrency"},
+            {"Weeping Essence of Torment", "StackableCurrency"},
+            {"Wailing Essence of Torment", "StackableCurrency"},
+            {"Screaming Essence of Torment", "StackableCurrency"},
+            {"Shrieking Essence of Torment", "StackableCurrency"},
+            {"Deafening Essence of Torment", "StackableCurrency"},
+            {"Whispering Essence of Woe", "StackableCurrency"},
+            {"Muttering Essence of Woe", "StackableCurrency"},
+            {"Weeping Essence of Woe", "StackableCurrency"},
+            {"Wailing Essence of Woe", "StackableCurrency"},
+            {"Screaming Essence of Woe", "StackableCurrency"},
+            {"Shrieking Essence of Woe", "StackableCurrency"},
+            {"Deafening Essence of Woe", "StackableCurrency"},
+            {"Weeping Essence of Wrath", "StackableCurrency"},
+            {"Wailing Essence of Wrath", "StackableCurrency"},
+            {"Screaming Essence of Wrath", "StackableCurrency"},
+            {"Shrieking Essence of Wrath", "StackableCurrency"},
+            {"Deafening Essence of Wrath", "StackableCurrency"},
+            {"Wailing Essence of Zeal", "StackableCurrency"},
+            {"Screaming Essence of Zeal", "StackableCurrency"},
+            {"Shrieking Essence of Zeal", "StackableCurrency"},
+            {"Deafening Essence of Zeal", "StackableCurrency"},
+            {"Glassblower's Bauble", "StackableCurrency"},
+            {"Gemcutter's Prism", "StackableCurrency"},
+            {"Time-light Scroll", "StackableCurrency"},
+            {"Deregulation Scroll", "StackableCurrency"},
+            {"Fragmentation Scroll", "StackableCurrency"},
+            {"Specularity Scroll", "StackableCurrency"},
+            {"Haemocombustion Scroll", "StackableCurrency"},
+            {"Electroshock Scroll", "StackableCurrency"},
+            {"Tailoring Orb", "StackableCurrency"},
+            {"Tempering Orb", "StackableCurrency"},
+            {"Scroll of Wisdom", "StackableCurrency"},
+            {"Scroll Fragment", "StackableCurrency"},
+            {"Imprint", "StackableCurrency"},
+            {"Eternal Orb", "StackableCurrency"},
+            {"Corrupt", "StackableCurrency"},
+            {"Vial of Sacrifice", "StackableCurrency"},
+            {"Vial of the Ghost", "StackableCurrency"},
+            {"Vial of Transcendence", "StackableCurrency"},
+            {"Vial of Fate", "StackableCurrency"},
+            {"Vial of Summoning", "StackableCurrency"},
+            {"Vial of the Ritual", "StackableCurrency"},
+            {"Vial of Consequence", "StackableCurrency"},
+            {"Vial of Awakening", "StackableCurrency"},
+            {"Vial of Dominance", "StackableCurrency"},
+            {"Bestiary Orb", "StackableCurrency"},
+            {"Imprinted Bestiary Orb", "StackableCurrency"},
+            {"Prophecy", "StackableCurrency"},
+            {"Abrasive Catalyst", "StackableCurrency"},
+            {"Intrinsic Catalyst", "StackableCurrency"},
+            {"Imbued Catalyst", "StackableCurrency"},
+            {"Unstable Catalyst", "StackableCurrency"},
+            {"Tempering Catalyst", "StackableCurrency"},
+            {"Turbulent Catalyst", "StackableCurrency"},
+            {"Noxious Catalyst", "StackableCurrency"},
+            {"Prismatic Catalyst", "StackableCurrency"},
+            {"Fertile Catalyst", "StackableCurrency"},
+            {"Accelerating Catalyst", "StackableCurrency"},
+            {"Enchant", "StackableCurrency"},
+            {"Timeless Eternal Empire Splinter", "StackableCurrency"},
+            {"Timeless Karui Splinter", "StackableCurrency"},
+            {"Timeless Maraketh Splinter", "StackableCurrency"},
+            {"Timeless Templar Splinter", "StackableCurrency"},
+            {"Timeless Vaal Splinter", "StackableCurrency"},
+            {"Cartographer's Chisel", "StackableCurrency"},
+            {"Divine Orb", "StackableCurrency"},
+            {"Orb of Regret", "StackableCurrency"},
+            {"Perandus Coin", "StackableCurrency"},
+            {"Portal Scroll", "StackableCurrency"},
+            {"Orb of Annulment", "StackableCurrency"},
+            {"Annulment Shard", "StackableCurrency"},
+            {"Blessed Orb", "StackableCurrency"},
+            {"Orb of Alteration", "StackableCurrency"},
+            {"Alteration Shard", "StackableCurrency"},
+            {"Orb of Horizons", "StackableCurrency"},
+            {"Horizon Shard", "StackableCurrency"},
+            {"Chaos Orb", "StackableCurrency"},
+            {"Chaos Shard", "StackableCurrency"},
+            {"Veiled Chaos Orb", "StackableCurrency"},
+            {"Prime Regrading Lens", "StackableCurrency"},
+            {"Chromatic Orb", "StackableCurrency"},
+            {"Orb of Fusing", "StackableCurrency"},
+            {"Jeweller's Orb", "StackableCurrency"},
+            {"Secondary Regrading Lens", "StackableCurrency"},
+            {"Ancient Orb", "StackableCurrency"},
+            {"Ancient Shard", "StackableCurrency"},
+            {"Unshaping Orb", "StackableCurrency"},
+            {"Albino Rhoa Feather", "StackableCurrency"},
+            {"Ritual Splinter", "StackableCurrency"},
+            {"Ritual Vessel", "StackableCurrency"},
+            {"Master Cartographer's Seal", "StackableCurrency"},
+            {"Apprentice Cartographer's Seal", "StackableCurrency"},
+            {"Journeyman Cartographer's Seal", "StackableCurrency"},
+            {"Silver Coin", "StackableCurrency"},
+            {"Engineer's Orb", "StackableCurrency"},
+            {"Infused Engineer's Orb", "StackableCurrency"},
+            {"Engineer's Shard", "StackableCurrency"},
+            {"Maven's Orb", "StackableCurrency"},
+            {"Regal Orb", "StackableCurrency"},
+            {"Regal Shard", "StackableCurrency"},
+            {"Harbinger's Orb", "StackableCurrency"},
+            {"Harbinger's Shard", "StackableCurrency"},
+            {"Orb of Chance", "StackableCurrency"},
+            {"Orb of Transmutation", "StackableCurrency"},
+            {"Transmutation Shard", "StackableCurrency"},
+            {"Orb of Alchemy", "StackableCurrency"},
+            {"Orb of Binding", "StackableCurrency"},
+            {"Binding Shard", "StackableCurrency"},
+            {"Alchemy Shard", "StackableCurrency"},
+            {"Blacksmith's Whetstone", "StackableCurrency"},
+            {"Clear Oil", "StackableCurrency"},
+            {"Opalescent Oil", "StackableCurrency"},
+            {"Silver Oil", "StackableCurrency"},
+            {"Golden Oil", "StackableCurrency"},
+            {"Sepia Oil", "StackableCurrency"},
+            {"Amber Oil", "StackableCurrency"},
+            {"Verdant Oil", "StackableCurrency"},
+            {"Teal Oil", "StackableCurrency"},
+            {"Azure Oil", "StackableCurrency"},
+            {"Indigo Oil", "StackableCurrency"},
+            {"Violet Oil", "StackableCurrency"},
+            {"Crimson Oil", "StackableCurrency"},
+            {"Black Oil", "StackableCurrency"},
+            {"Tainted Oil", "StackableCurrency"},
+            {"A Dab of Ink", "DivinationCard"},
+            {"A Familiar Call", "DivinationCard"},
+            {"A Modest Request", "DivinationCard"},
+            {"A Mother's Parting Gift", "DivinationCard"},
+            {"A Note in the Wind", "DivinationCard"},
+            {"Abandoned Wealth", "DivinationCard"},
+            {"Acclimatisation", "DivinationCard"},
+            {"Akil's Prophecy", "DivinationCard"},
+            {"Alluring Bounty", "DivinationCard"},
+            {"Alone in the Darkness", "DivinationCard"},
+            {"Anarchy's Price", "DivinationCard"},
+            {"Arrogance of the Vaal", "DivinationCard"},
+            {"Assassin's Favour", "DivinationCard"},
+            {"Atziri's Arsenal", "DivinationCard"},
+            {"Audacity", "DivinationCard"},
+            {"Azyran's Reward", "DivinationCard"},
+            {"Baited Expectations", "DivinationCard"},
+            {"Beauty Through Death", "DivinationCard"},
+            {"Birth of the Three", "DivinationCard"},
+            {"Blessing of God", "DivinationCard"},
+            {"Blind Venture", "DivinationCard"},
+            {"Boon of Justice", "DivinationCard"},
+            {"Boon of the First Ones", "DivinationCard"},
+            {"Boundless Realms", "DivinationCard"},
+            {"Bowyer's Dream", "DivinationCard"},
+            {"Broken Promises", "DivinationCard"},
+            {"Brotherhood in Exile", "DivinationCard"},
+            {"Brother's Stash", "DivinationCard"},
+            {"Brush, Paint and Palette", "DivinationCard"},
+            {"Buried Treasure", "DivinationCard"},
+            {"Burning Blood", "DivinationCard"},
+            {"Call to the First Ones", "DivinationCard"},
+            {"Cameria's Cut", "DivinationCard"},
+            {"Cartographer's Delight", "DivinationCard"},
+            {"Chaotic Disposition", "DivinationCard"},
+            {"Council of Cats", "DivinationCard"},
+            {"Coveted Possession", "DivinationCard"},
+            {"Cursed Words", "DivinationCard"},
+            {"Dark Dreams", "DivinationCard"},
+            {"Dark Temptation", "DivinationCard"},
+            {"Deadly Joy", "DivinationCard"},
+            {"Death", "DivinationCard"},
+            {"Deathly Designs", "DivinationCard"},
+            {"Stacked Deck", "StackableCurrency"},
+            {"Demigod's Wager", "DivinationCard"},
+            {"Desecrated Virtue", "DivinationCard"},
+            {"Destined to Crumble", "DivinationCard"},
+            {"Dialla's Subjugation", "DivinationCard"},
+            {"Divine Justice", "DivinationCard"},
+            {"Doedre's Madness", "DivinationCard"},
+            {"Draped in Dreams", "DivinationCard"},
+            {"Dying Anguish", "DivinationCard"},
+            {"Dying Light", "DivinationCard"},
+            {"Earth Drinker", "DivinationCard"},
+            {"Echoes of Love", "DivinationCard"},
+            {"Emperor of Purity", "DivinationCard"},
+            {"Emperor's Luck", "DivinationCard"},
+            {"Etched in Blood", "DivinationCard"},
+            {"Fateful Meeting", "DivinationCard"},
+            {"Forbidden Power", "DivinationCard"},
+            {"Friendship", "DivinationCard"},
+            {"Gemcutter's Promise", "DivinationCard"},
+            {"Gift of Asenath", "DivinationCard"},
+            {"Gift of the Gemling Queen", "DivinationCard"},
+            {"Glimmer of Hope", "DivinationCard"},
+            {"Grave Knowledge", "DivinationCard"},
+            {"Harmony of Souls", "DivinationCard"},
+            {"Haunting Shadows", "DivinationCard"},
+            {"Her Mask", "DivinationCard"},
+            {"Heterochromia", "DivinationCard"},
+            {"Hope", "DivinationCard"},
+            {"House of Mirrors", "DivinationCard"},
+            {"Hubris", "DivinationCard"},
+            {"Humility", "DivinationCard"},
+            {"Hunter's Resolve", "DivinationCard"},
+            {"Hunter's Reward", "DivinationCard"},
+            {"Immortal Resolve", "DivinationCard"},
+            {"Imperial Legacy", "DivinationCard"},
+            {"Jack in the Box", "DivinationCard"},
+            {"Keeper's Corruption", "DivinationCard"},
+            {"Lantador's Lost Love", "DivinationCard"},
+            {"Last Hope", "DivinationCard"},
+            {"Left to Fate", "DivinationCard"},
+            {"Light and Truth", "DivinationCard"},
+            {"Lingering Remnants", "DivinationCard"},
+            {"Lost Worlds", "DivinationCard"},
+            {"Love Through Ice", "DivinationCard"},
+            {"Loyalty", "DivinationCard"},
+            {"Luck of the Vaal", "DivinationCard"},
+            {"Lucky Connections", "DivinationCard"},
+            {"Lucky Deck", "DivinationCard"},
+            {"Luminous Trove", "DivinationCard"},
+            {"Lysah's Respite", "DivinationCard"},
+            {"Mawr Blaidd", "DivinationCard"},
+            {"Merciless Armament", "DivinationCard"},
+            {"Might is Right", "DivinationCard"},
+            {"Mitts", "DivinationCard"},
+            {"Monochrome", "DivinationCard"},
+            {"More is Never Enough", "DivinationCard"},
+            {"No Traces", "DivinationCard"},
+            {"Nook's Crown", "DivinationCard"},
+            {"Peaceful Moments", "DivinationCard"},
+            {"Perfection", "DivinationCard"},
+            {"Prejudice", "DivinationCard"},
+            {"Pride Before the Fall", "DivinationCard"},
+            {"Pride of the First Ones", "DivinationCard"},
+            {"Prometheus' Armoury", "DivinationCard"},
+            {"Prosperity", "DivinationCard"},
+            {"Rain of Chaos", "DivinationCard"},
+            {"Rain Tempter", "DivinationCard"},
+            {"Rats", "DivinationCard"},
+            {"Rebirth", "DivinationCard"},
+            {"Reckless Ambition", "DivinationCard"},
+            {"Remembrance", "DivinationCard"},
+            {"Sambodhi's Vow", "DivinationCard"},
+            {"Sambodhi's Wisdom", "DivinationCard"},
+            {"Scholar of the Seas", "DivinationCard"},
+            {"Seven Years Bad Luck", "DivinationCard"},
+            {"Shard of Fate", "DivinationCard"},
+            {"Society's Remorse", "DivinationCard"},
+            {"Squandered Prosperity", "DivinationCard"},
+            {"Struck by Lightning", "DivinationCard"},
+            {"Succor of the Sinless", "DivinationCard"},
+            {"The Academic", "DivinationCard"},
+            {"The Admirer", "DivinationCard"},
+            {"The Adventuring Spirit", "DivinationCard"},
+            {"The Aesthete", "DivinationCard"},
+            {"The Archmage's Right Hand", "DivinationCard"},
+            {"The Arena Champion", "DivinationCard"},
+            {"The Army of Blood", "DivinationCard"},
+            {"The Artist", "DivinationCard"},
+            {"The Astromancer", "DivinationCard"},
+            {"The Avenger", "DivinationCard"},
+            {"The Awakened", "DivinationCard"},
+            {"The Bargain", "DivinationCard"},
+            {"The Battle Born", "DivinationCard"},
+            {"The Bear Woman", "DivinationCard"},
+            {"The Beast", "DivinationCard"},
+            {"The Betrayal", "DivinationCard"},
+            {"The Bitter Blossom", "DivinationCard"},
+            {"The Blazing Fire", "DivinationCard"},
+            {"The Body", "DivinationCard"},
+            {"The Bones", "DivinationCard"},
+            {"The Breach", "DivinationCard"},
+            {"The Brittle Emperor", "DivinationCard"},
+            {"The Cache", "DivinationCard"},
+            {"The Cacophony", "DivinationCard"},
+            {"The Calling", "DivinationCard"},
+            {"The Card Sharp", "DivinationCard"},
+            {"The Carrion Crow", "DivinationCard"},
+            {"The Cartographer", "DivinationCard"},
+            {"The Cataclysm", "DivinationCard"},
+            {"The Catalyst", "DivinationCard"},
+            {"The Celestial Justicar", "DivinationCard"},
+            {"The Celestial Stone", "DivinationCard"},
+            {"The Chains that Bind", "DivinationCard"},
+            {"The Cheater", "DivinationCard"},
+            {"The Chosen", "DivinationCard"},
+            {"The Coming Storm", "DivinationCard"},
+            {"The Conduit", "DivinationCard"},
+            {"The Craving", "DivinationCard"},
+            {"The Cursed King", "DivinationCard"},
+            {"The Damned", "DivinationCard"},
+            {"The Dapper Prodigy", "DivinationCard"},
+            {"The Dark Mage", "DivinationCard"},
+            {"The Darkest Dream", "DivinationCard"},
+            {"The Deal", "DivinationCard"},
+            {"The Deceiver", "DivinationCard"},
+            {"The Deep Ones", "DivinationCard"},
+            {"The Demon", "DivinationCard"},
+            {"The Demoness", "DivinationCard"},
+            {"The Devastator", "DivinationCard"},
+            {"The Doctor", "DivinationCard"},
+            {"The Doppelganger", "DivinationCard"},
+            {"The Dragon", "DivinationCard"},
+            {"The Dragon's Heart", "DivinationCard"},
+            {"The Dreamer", "DivinationCard"},
+            {"The Dreamland", "DivinationCard"},
+            {"The Drunken Aristocrat", "DivinationCard"},
+            {"The Easy Stroll", "DivinationCard"},
+            {"The Eldritch Decay", "DivinationCard"},
+            {"The Encroaching Darkness", "DivinationCard"},
+            {"The Endless Darkness", "DivinationCard"},
+            {"The Endurance", "DivinationCard"},
+            {"The Enlightened", "DivinationCard"},
+            {"The Enthusiasts", "DivinationCard"},
+            {"The Escape", "DivinationCard"},
+            {"The Eternal War", "DivinationCard"},
+            {"The Ethereal", "DivinationCard"},
+            {"The Explorer", "DivinationCard"},
+            {"The Eye of Terror", "DivinationCard"},
+            {"The Eye of the Dragon", "DivinationCard"},
+            {"The Fathomless Depths", "DivinationCard"},
+            {"The Feast", "DivinationCard"},
+            {"The Fiend", "DivinationCard"},
+            {"The Fishmonger", "DivinationCard"},
+            {"The Fletcher", "DivinationCard"},
+            {"The Flora's Gift", "DivinationCard"},
+            {"The Fool", "DivinationCard"},
+            {"The Formless Sea", "DivinationCard"},
+            {"The Forsaken", "DivinationCard"},
+            {"The Fox", "DivinationCard"},
+            {"The Gambler", "DivinationCard"},
+            {"The Garish Power", "DivinationCard"},
+            {"The Gemcutter", "DivinationCard"},
+            {"The Gentleman", "DivinationCard"},
+            {"The Gladiator", "DivinationCard"},
+            {"The Golden Era", "DivinationCard"},
+            {"The Greatest Intentions", "DivinationCard"},
+            {"The Gulf", "DivinationCard"},
+            {"The Hale Heart", "DivinationCard"},
+            {"The Harvester", "DivinationCard"},
+            {"The Hermit", "DivinationCard"},
+            {"The Heroic Shot", "DivinationCard"},
+            {"The Hive of Knowledge", "DivinationCard"},
+            {"The Hoarder", "DivinationCard"},
+            {"The Hook", "DivinationCard"},
+            {"The Hunger", "DivinationCard"},
+            {"The Immortal", "DivinationCard"},
+            {"The Incantation", "DivinationCard"},
+            {"The Innocent", "DivinationCard"},
+            {"The Inoculated", "DivinationCard"},
+            {"The Insatiable", "DivinationCard"},
+            {"The Inventor", "DivinationCard"},
+            {"The Iron Bard", "DivinationCard"},
+            {"The Jester", "DivinationCard"},
+            {"The Jeweller's Boon", "DivinationCard"},
+            {"The Journalist", "DivinationCard"},
+            {"The Journey", "DivinationCard"},
+            {"The King's Blade", "DivinationCard"},
+            {"The King's Heart", "DivinationCard"},
+            {"The Landing", "DivinationCard"},
+            {"The Last One Standing", "DivinationCard"},
+            {"The Lich", "DivinationCard"},
+            {"The Life Thief", "DivinationCard"},
+            {"The Lion", "DivinationCard"},
+            {"The Long Con", "DivinationCard"},
+            {"The Long Watch", "DivinationCard"},
+            {"The Lord in Black", "DivinationCard"},
+            {"The Lord of Celebration", "DivinationCard"},
+            {"The Lover", "DivinationCard"},
+            {"The Lunaris Priestess", "DivinationCard"},
+            {"The Mad King", "DivinationCard"},
+            {"The Master", "DivinationCard"},
+            {"The Master Artisan", "DivinationCard"},
+            {"The Mayor", "DivinationCard"},
+            {"The Mercenary", "DivinationCard"},
+            {"The Messenger", "DivinationCard"},
+            {"The Metalsmith's Gift", "DivinationCard"},
+            {"The Mountain", "DivinationCard"},
+            {"The Nurse", "DivinationCard"},
+            {"The Oath", "DivinationCard"},
+            {"The Obscured", "DivinationCard"},
+            {"The Offering", "DivinationCard"},
+            {"The Old Man", "DivinationCard"},
+            {"The One With All", "DivinationCard"},
+            {"The Opulent", "DivinationCard"},
+            {"The Pack Leader", "DivinationCard"},
+            {"The Pact", "DivinationCard"},
+            {"The Patient", "DivinationCard"},
+            {"The Penitent", "DivinationCard"},
+            {"The Poet", "DivinationCard"},
+            {"The Polymath", "DivinationCard"},
+            {"The Porcupine", "DivinationCard"},
+            {"The Price of Loyalty", "DivinationCard"},
+            {"The Price of Protection", "DivinationCard"},
+            {"The Primordial", "DivinationCard"},
+            {"The Professor", "DivinationCard"},
+            {"The Progeny of Lunaris", "DivinationCard"},
+            {"The Puzzle", "DivinationCard"},
+            {"The Queen", "DivinationCard"},
+            {"The Rabid Rhoa", "DivinationCard"},
+            {"The Realm", "DivinationCard"},
+            {"The Risk", "DivinationCard"},
+            {"The Rite of Elements", "DivinationCard"},
+            {"The Road to Power", "DivinationCard"},
+            {"The Ruthless Ceinture", "DivinationCard"},
+            {"The Sacrifice", "DivinationCard"},
+            {"The Saint's Treasure", "DivinationCard"},
+            {"The Samurai's Eye", "DivinationCard"},
+            {"The Scarred Meadow", "DivinationCard"},
+            {"The Scavenger", "DivinationCard"},
+            {"The Scholar", "DivinationCard"},
+            {"The Seeker", "DivinationCard"},
+            {"The Sephirot", "DivinationCard"},
+            {"The Shortcut", "DivinationCard"},
+            {"The Side Quest", "DivinationCard"},
+            {"The Sigil", "DivinationCard"},
+            {"The Siren", "DivinationCard"},
+            {"The Skeleton", "DivinationCard"},
+            {"The Soul", "DivinationCard"},
+            {"The Spark and the Flame", "DivinationCard"},
+            {"The Spoiled Prince", "DivinationCard"},
+            {"The Standoff", "DivinationCard"},
+            {"The Stormcaller", "DivinationCard"},
+            {"The Strategist", "DivinationCard"},
+            {"The Summoner", "DivinationCard"},
+            {"The Sun", "DivinationCard"},
+            {"The Surgeon", "DivinationCard"},
+            {"The Surveyor", "DivinationCard"},
+            {"The Survivalist", "DivinationCard"},
+            {"The Sustenance", "DivinationCard"},
+            {"The Sword King's Salute", "DivinationCard"},
+            {"The Thaumaturgist", "DivinationCard"},
+            {"The Throne", "DivinationCard"},
+            {"The Tinkerer's Table", "DivinationCard"},
+            {"The Tower", "DivinationCard"},
+            {"The Traitor", "DivinationCard"},
+            {"The Trial", "DivinationCard"},
+            {"The Tumbleweed", "DivinationCard"},
+            {"The Twilight Moon", "DivinationCard"},
+            {"The Twins", "DivinationCard"},
+            {"The Tyrant", "DivinationCard"},
+            {"The Undaunted", "DivinationCard"},
+            {"The Undisputed", "DivinationCard"},
+            {"The Unexpected Prize", "DivinationCard"},
+            {"The Union", "DivinationCard"},
+            {"The Valkyrie", "DivinationCard"},
+            {"The Valley of Steel Boxes", "DivinationCard"},
+            {"The Vast", "DivinationCard"},
+            {"The Visionary", "DivinationCard"},
+            {"The Void", "DivinationCard"},
+            {"The Warden", "DivinationCard"},
+            {"The Warlord", "DivinationCard"},
+            {"The Watcher", "DivinationCard"},
+            {"The Web", "DivinationCard"},
+            {"The White Knight", "DivinationCard"},
+            {"The Whiteout", "DivinationCard"},
+            {"The Wilted Rose", "DivinationCard"},
+            {"The Wind", "DivinationCard"},
+            {"The Witch", "DivinationCard"},
+            {"The Wolf", "DivinationCard"},
+            {"The Wolf's Legacy", "DivinationCard"},
+            {"The Wolf's Shadow", "DivinationCard"},
+            {"The Wolven King's Bite", "DivinationCard"},
+            {"The Wolverine", "DivinationCard"},
+            {"The World Eater", "DivinationCard"},
+            {"The Wrath", "DivinationCard"},
+            {"The Wretched", "DivinationCard"},
+            {"Thirst for Knowledge", "DivinationCard"},
+            {"Three Faces in the Dark", "DivinationCard"},
+            {"Three Voices", "DivinationCard"},
+            {"Thunderous Skies", "DivinationCard"},
+            {"Time-Lost Relic", "DivinationCard"},
+            {"Tranquillity", "DivinationCard"},
+            {"Treasure Hunter", "DivinationCard"},
+            {"Triskaidekaphobia", "DivinationCard"},
+            {"Turn the Other Cheek", "DivinationCard"},
+            {"Unchained", "DivinationCard"},
+            {"Underground Forest", "DivinationCard"},
+            {"Unrequited Love", "DivinationCard"},
+            {"Vanity", "DivinationCard"},
+            {"Vile Power", "DivinationCard"},
+            {"Vinia's Token", "DivinationCard"},
+            {"Void of the Elements", "DivinationCard"},
+            {"Volatile Power", "DivinationCard"},
+            {"Wealth and Power", "DivinationCard"},
+            {"Winter's Embrace", "DivinationCard"},
+            {"Small Hybrid Flask", "HybridFlask"},
+            {"Medium Hybrid Flask", "HybridFlask"},
+            {"Large Hybrid Flask", "HybridFlask"},
+            {"Colossal Hybrid Flask", "HybridFlask"},
+            {"Sacred Hybrid Flask", "HybridFlask"},
+            {"Hallowed Hybrid Flask", "HybridFlask"},
+            {"Small Life Flask", "LifeFlask"},
+            {"Sanctified Life Flask", "LifeFlask"},
+            {"Divine Life Flask", "LifeFlask"},
+            {"Eternal Life Flask", "LifeFlask"},
+            {"Medium Life Flask", "LifeFlask"},
+            {"Large Life Flask", "LifeFlask"},
+            {"Greater Life Flask", "LifeFlask"},
+            {"Grand Life Flask", "LifeFlask"},
+            {"Giant Life Flask", "LifeFlask"},
+            {"Colossal Life Flask", "LifeFlask"},
+            {"Sacred Life Flask", "LifeFlask"},
+            {"Hallowed Life Flask", "LifeFlask"},
+            {"Small Mana Flask", "ManaFlask"},
+            {"Sanctified Mana Flask", "ManaFlask"},
+            {"Divine Mana Flask", "ManaFlask"},
+            {"Eternal Mana Flask", "ManaFlask"},
+            {"Medium Mana Flask", "ManaFlask"},
+            {"Large Mana Flask", "ManaFlask"},
+            {"Greater Mana Flask", "ManaFlask"},
+            {"Grand Mana Flask", "ManaFlask"},
+            {"Giant Mana Flask", "ManaFlask"},
+            {"Colossal Mana Flask", "ManaFlask"},
+            {"Sacred Mana Flask", "ManaFlask"},
+            {"Hallowed Mana Flask", "ManaFlask"},
+            {"Diamond Flask", "UtilityFlaskCritical"},
+            {"Basalt Flask", "UtilityFlask"},
+            {"Aquamarine Flask", "UtilityFlask"},
+            {"Stibnite Flask", "UtilityFlask"},
+            {"Sulphur Flask", "UtilityFlask"},
+            {"Silver Flask", "UtilityFlask"},
+            {"Bismuth Flask", "UtilityFlask"},
+            {"Ruby Flask", "UtilityFlask"},
+            {"Sapphire Flask", "UtilityFlask"},
+            {"Topaz Flask", "UtilityFlask"},
+            {"Granite Flask", "UtilityFlask"},
+            {"Quicksilver Flask", "UtilityFlask"},
+            {"Amethyst Flask", "UtilityFlask"},
+            {"Quartz Flask", "UtilityFlask"},
+            {"Jade Flask", "UtilityFlask"},
+            {"Infernal Cry", "Active Skill Gem"},
+            {"Ancestral Cry", "Active Skill Gem"},
+            {"Ancestral Warchief", "Active Skill Gem"},
+            {"Anger", "Active Skill Gem"},
+            {"Animate Guardian", "Active Skill Gem"},
+            {"Animate Weapon", "Active Skill Gem"},
+            {"Arc", "Active Skill Gem"},
+            {"Arcane Cloak", "Active Skill Gem"},
+            {"Arcanist Brand", "Active Skill Gem"},
+            {"Arctic Armour", "Active Skill Gem"},
+            {"Creeping Frost", "Active Skill Gem"},
+            {"Armageddon Brand", "Active Skill Gem"},
+            {"Artillery Ballista", "Active Skill Gem"},
+            {"Backstab", "Active Skill Gem"},
+            {"Ball Lightning", "Active Skill Gem"},
+            {"Barrage", "Active Skill Gem"},
+            {"Bear Trap", "Active Skill Gem"},
+            {"Berserk", "Active Skill Gem"},
+            {"Blade Blast", "Active Skill Gem"},
+            {"Blade Trap", "Active Skill Gem"},
+            {"Blade Vortex", "Active Skill Gem"},
+            {"Bladefall", "Active Skill Gem"},
+            {"Bladestorm", "Active Skill Gem"},
+            {"Blast Rain", "Active Skill Gem"},
+            {"Blazing Salvo", "Active Skill Gem"},
+            {"Blight", "Active Skill Gem"},
+            {"Blinding Aura", "Active Skill Gem"},
+            {"Blink Arrow", "Active Skill Gem"},
+            {"[UNUSED] Blitz", "Active Skill Gem"},
+            {"Blood and Sand", "Active Skill Gem"},
+            {"Blood Rage", "Active Skill Gem"},
+            {"[UNUSED] Blood Whirl", "Active Skill Gem"},
+            {"Reap", "Active Skill Gem"},
+            {"[UNUSED] Bone Armour", "Active Skill Gem"},
+            {"Unearth", "Active Skill Gem"},
+            {"Bone Offering", "Active Skill Gem"},
+            {"Burning Arrow", "Active Skill Gem"},
+            {"Capture Monster", "Active Skill Gem"},
+            {"Chain Hook", "Active Skill Gem"},
+            {"Blade Flurry", "Active Skill Gem"},
+            {"Charged Dash", "Active Skill Gem"},
+            {"Clarity", "Active Skill Gem"},
+            {"Cleave", "Active Skill Gem"},
+            {"Cobra Lash", "Active Skill Gem"},
+            {"[UNUSED] Coiling Assault", "Active Skill Gem"},
+            {"Purity of Ice", "Active Skill Gem"},
+            {"Cold Snap", "Active Skill Gem"},
+            {"...", "Active Skill Gem"},
+            {"Conductivity", "Active Skill Gem"},
+            {"Consecrated Path", "Active Skill Gem"},
+            {"Contagion", "Active Skill Gem"},
+            {"Conversion Trap", "Active Skill Gem"},
+            {"Convocation", "Active Skill Gem"},
+            {"Cremation", "Active Skill Gem"},
+            {"Bodyswap", "Active Skill Gem"},
+            {"Corrupting Fever", "Active Skill Gem"},
+            {"Crackling Lance", "Active Skill Gem"},
+            {"Assassin's Mark", "Active Skill Gem"},
+            {"Cyclone", "Active Skill Gem"},
+            {"Damage Infusion", "Active Skill Gem"},
+            {"Malevolence", "Active Skill Gem"},
+            {"Dark Pact", "Active Skill Gem"},
+            {"Bane", "Active Skill Gem"},
+            {"Dash", "Active Skill Gem"},
+            {"Death Aura", "Active Skill Gem"},
+            {"Decoy Totem", "Active Skill Gem"},
+            {"Desecrate", "Active Skill Gem"},
+            {"Determination", "Active Skill Gem"},
+            {"Detonate Dead", "Active Skill Gem"},
+            {"Detonate Mines", "Active Skill Gem"},
+            {"Devouring Totem", "Active Skill Gem"},
+            {"Discharge", "Active Skill Gem"},
+            {"Discipline", "Active Skill Gem"},
+            {"Discorectangle Slam", "Active Skill Gem"},
+            {"Divine Ire", "Active Skill Gem"},
+            {"Dominating Blow", "Active Skill Gem"},
+            {"Double Strike", "Active Skill Gem"},
+            {"Dread Banner", "Active Skill Gem"},
+            {"Dual Strike", "Active Skill Gem"},
+            {"Earthquake", "Active Skill Gem"},
+            {"Earthshatter", "Active Skill Gem"},
+            {"Elemental Hit", "Active Skill Gem"},
+            {"Elemental Projectiles", "Active Skill Gem"},
+            {"Elemental Weakness", "Active Skill Gem"},
+            {"Enduring Cry", "Active Skill Gem"},
+            {"Enfeeble", "Active Skill Gem"},
+            {"Ensnaring Arrow", "Active Skill Gem"},
+            {"Envy", "Active Skill Gem"},
+            {"Essence Drain", "Active Skill Gem"},
+            {"Ethereal Knives", "Active Skill Gem"},
+            {"Explosive Arrow", "Active Skill Gem"},
+            {"Exsanguinate", "Active Skill Gem"},
+            {"Scorching Ray", "Active Skill Gem"},
+            {"Pyroclast Mine", "Active Skill Gem"},
+            {"Purity of Fire", "Active Skill Gem"},
+            {"Fire Trap", "Active Skill Gem"},
+            {"Fire Weapon", "Active Skill Gem"},
+            {"Fireball", "Active Skill Gem"},
+            {"Firestorm", "Active Skill Gem"},
+            {"Flame Dash", "Active Skill Gem"},
+            {"Holy Flame Totem", "Active Skill Gem"},
+            {"Flame Wall", "Active Skill Gem"},
+            {"Flame Surge", "Active Skill Gem"},
+            {"Flameblast", "Active Skill Gem"},
+            {"Flamethrower Trap", "Active Skill Gem"},
+            {"Flammability", "Active Skill Gem"},
+            {"Flammable Shot", "Active Skill Gem"},
+            {"Flesh and Stone", "Active Skill Gem"},
+            {"Flesh Offering", "Active Skill Gem"},
+            {"Flicker Strike", "Active Skill Gem"},
+            {"Icicle Mine", "Active Skill Gem"},
+            {"Freezing Pulse", "Active Skill Gem"},
+            {"Frenzy", "Active Skill Gem"},
+            {"Frost Blades", "Active Skill Gem"},
+            {"Frostbolt", "Active Skill Gem"},
+            {"Vortex", "Active Skill Gem"},
+            {"Frost Bomb", "Active Skill Gem"},
+            {"Frost Shield", "Active Skill Gem"},
+            {"Frost Wall", "Active Skill Gem"},
+            {"Frostbite", "Active Skill Gem"},
+            {"Frostblink", "Active Skill Gem"},
+            {"General's Cry", "Active Skill Gem"},
+            {"Glacial Cascade", "Active Skill Gem"},
+            {"Glacial Hammer", "Active Skill Gem"},
+            {"Gluttony of Elements", "Active Skill Gem"},
+            {"Grace", "Active Skill Gem"},
+            {"Ground Slam", "Active Skill Gem"},
+            {"Haste", "Active Skill Gem"},
+            {"Hatred", "Active Skill Gem"},
+            {"Heavy Strike", "Active Skill Gem"},
+            {"Herald of Agony", "Active Skill Gem"},
+            {"Herald of Ash", "Active Skill Gem"},
+            {"Herald of Blood", "Active Skill Gem"},
+            {"Herald of Ice", "Active Skill Gem"},
+            {"Herald of Purity", "Active Skill Gem"},
+            {"Herald of Thunder", "Active Skill Gem"},
+            {"Hexblast", "Active Skill Gem"},
+            {"Hydrosphere", "Active Skill Gem"},
+            {"Ice Crash", "Active Skill Gem"},
+            {"Ice Nova", "Active Skill Gem"},
+            {"Ice Shot", "Active Skill Gem"},
+            {"Siphoning Trap", "Active Skill Gem"},
+            {"Ice Spear", "Active Skill Gem"},
+            {"Ice Trap", "Active Skill Gem"},
+            {"Icefire", "Active Skill Gem"},
+            {"Ignite", "Active Skill Gem"},
+            {"Immortal Call", "Active Skill Gem"},
+            {"Incinerate", "Active Skill Gem"},
+            {"Infernal Blow", "Active Skill Gem"},
+            {"Wildfire", "Active Skill Gem"},
+            {"Infernal Sweep", "Active Skill Gem"},
+            {"Intimidating Cry", "Active Skill Gem"},
+            {"Kinetic Blast", "Active Skill Gem"},
+            {"Kinetic Bolt", "Active Skill Gem"},
+            {"Lacerate", "Active Skill Gem"},
+            {"Lancing Steel", "Active Skill Gem"},
+            {"Leap Slam", "Active Skill Gem"},
+            {"Lightning Arrow", "Active Skill Gem"},
+            {"Lightning Channel", "Active Skill Gem"},
+            {"Lightning Circle", "Active Skill Gem"},
+            {"Purity of Lightning", "Active Skill Gem"},
+            {"Lightning Strike", "Active Skill Gem"},
+            {"Lightning Tendrils", "Active Skill Gem"},
+            {"Lightning Spire Trap", "Active Skill Gem"},
+            {"Lightning Trap", "Active Skill Gem"},
+            {"Lightning Warp", "Active Skill Gem"},
+            {"Rolling Magma", "Active Skill Gem"},
+            {"Ancestral Protector", "Active Skill Gem"},
+            {"Mirror Arrow", "Active Skill Gem"},
+            {"Molten Shell", "Active Skill Gem"},
+            {"Molten Strike", "Active Skill Gem"},
+            {"New Blade Vortex", "Active Skill Gem"},
+            {"Phase Run", "Active Skill Gem"},
+            {"NewPunishment", "Active Skill Gem"},
+            {"New Shock Nova", "Active Skill Gem"},
+            {"Vulnerability", "Active Skill Gem"},
+            {"Orb of Storms", "Active Skill Gem"},
+            {"Penance Brand", "Active Skill Gem"},
+            {"Perforate", "Active Skill Gem"},
+            {"Pestilent Strike", "Active Skill Gem"},
+            {"Petrified Blood", "Active Skill Gem"},
+            {"Seismic Trap", "Active Skill Gem"},
+            {"Plague Bearer", "Active Skill Gem"},
+            {"Poacher's Mark", "Active Skill Gem"},
+            {"Caustic Arrow", "Active Skill Gem"},
+            {"Portal", "Active Skill Gem"},
+            {"Power Siphon", "Active Skill Gem"},
+            {"Precision", "Active Skill Gem"},
+            {"Pride", "Active Skill Gem"},
+            {"[UNUSED] Projectile Portal", "Active Skill Gem"},
+            {"Sniper's Mark", "Active Skill Gem"},
+            {"Puncture", "Active Skill Gem"},
+            {"Punishment", "Active Skill Gem"},
+            {"Wave of Conviction", "Active Skill Gem"},
+            {"Purity of Elements", "Active Skill Gem"},
+            {"[UNUSED] Quick Block", "Active Skill Gem"},
+            {"Rain of Arrows", "Active Skill Gem"},
+            {"Raise Spectre", "Active Skill Gem"},
+            {"Raise Zombie", "Active Skill Gem"},
+            {"Rallying Cry", "Active Skill Gem"},
+            {"Reave", "Active Skill Gem"},
+            {"Brand Recall", "Active Skill Gem"},
+            {"Reckoning", "Active Skill Gem"},
+            {"Rejuvenation Totem", "Active Skill Gem"},
+            {"Rending Steel", "Active Skill Gem"},
+            {"[UNUSED] Replicate", "Active Skill Gem"},
+            {"Righteous Fire", "Active Skill Gem"},
+            {"Righteous Lightning", "Active Skill Gem"},
+            {"Riposte", "Active Skill Gem"},
+            {"Riptide", "Active Skill Gem"},
+            {"Purifying Flame", "Active Skill Gem"},
+            {"Scourge Arrow", "Active Skill Gem"},
+            {"Searing Bond", "Active Skill Gem"},
+            {"Seismic Cry", "Active Skill Gem"},
+            {"[UNUSED] Serpent Strike", "Active Skill Gem"},
+            {"Shadow Blades", "Active Skill Gem"},
+            {"Shattering Steel", "Active Skill Gem"},
+            {"Shield Charge", "Active Skill Gem"},
+            {"Shock Nova", "Active Skill Gem"},
+            {"Shockwave Totem", "Active Skill Gem"},
+            {"Shrapnel Ballista", "Active Skill Gem"},
+            {"Galvanic Arrow", "Active Skill Gem"},
+            {"Explosive Trap", "Active Skill Gem"},
+            {"Siege Ballista", "Active Skill Gem"},
+            {"Sigil of Power", "Active Skill Gem"},
+            {"Ancestral Blademaster", "Active Skill Gem"},
+            {"[UNUSED] Slice And Dice", "Active Skill Gem"},
+            {"Smite", "Active Skill Gem"},
+            {"Smoke Mine", "Active Skill Gem"},
+            {"Snipe", "Active Skill Gem"},
+            {"Soulrend", "Active Skill Gem"},
+            {"Spark", "Active Skill Gem"},
+            {"Spectral Spinning Weapon", "Active Skill Gem"},
+            {"Zealotry", "Active Skill Gem"},
+            {"Spellslinger", "Active Skill Gem"},
+            {"Spirit Offering", "Active Skill Gem"},
+            {"Split Arrow", "Active Skill Gem"},
+            {"Splitting Steel", "Active Skill Gem"},
+            {"Static Strike", "Active Skill Gem"},
+            {"Static Tether", "Active Skill Gem"},
+            {"Steelskin", "Active Skill Gem"},
+            {"Stormbind", "Active Skill Gem"},
+            {"Storm Brand", "Active Skill Gem"},
+            {"Storm Burst", "Active Skill Gem"},
+            {"Storm Call", "Active Skill Gem"},
+            {"Stormblast Mine", "Active Skill Gem"},
+            {"Summon Carrion Golem", "Active Skill Gem"},
+            {"Summon Chaos Golem", "Active Skill Gem"},
+            {"Summon Flame Golem", "Active Skill Gem"},
+            {"Summon Ice Golem", "Active Skill Gem"},
+            {"Summon Lightning Golem", "Active Skill Gem"},
+            {"Summon Raging Spirit", "Active Skill Gem"},
+            {"Summon Holy Relic", "Active Skill Gem"},
+            {"Summon Stone Golem", "Active Skill Gem"},
+            {"Summon Skeletons", "Active Skill Gem"},
+            {"Summon Skeletons Channelled", "Active Skill Gem"},
+            {"Summon Skitterbots", "Active Skill Gem"},
+            {"Sunder", "Active Skill Gem"},
+            {"Sweep", "Active Skill Gem"},
+            {"Tectonic Slam", "Active Skill Gem"},
+            {"Tempest Shield", "Active Skill Gem"},
+            {"Temporal Chains", "Active Skill Gem"},
+            {"Spectral Shield Throw", "Active Skill Gem"},
+            {"Spectral Throw", "Active Skill Gem"},
+            {"Tornado Shot", "Active Skill Gem"},
+            {"Touch of God", "Active Skill Gem"},
+            {"Toxic Rain", "Active Skill Gem"},
+            {"Vaal Ancestral Warchief", "Active Skill Gem"},
+            {"Vaal Arc", "Active Skill Gem"},
+            {"Vaal Blade Vortex", "Active Skill Gem"},
+            {"Vaal Blight", "Active Skill Gem"},
+            {"Vaal Burning Arrow", "Active Skill Gem"},
+            {"Vaal Clarity", "Active Skill Gem"},
+            {"Vaal Impurity of Ice", "Active Skill Gem"},
+            {"Vaal Cold Snap", "Active Skill Gem"},
+            {"Vaal Cyclone", "Active Skill Gem"},
+            {"Vaal Detonate Dead", "Active Skill Gem"},
+            {"Vaal Discipline", "Active Skill Gem"},
+            {"Vaal Double Strike", "Active Skill Gem"},
+            {"Vaal Earthquake", "Active Skill Gem"},
+            {"Vaal Impurity of Fire", "Active Skill Gem"},
+            {"Vaal FireTrap", "Active Skill Gem"},
+            {"Vaal Fireball", "Active Skill Gem"},
+            {"Vaal Flameblast", "Active Skill Gem"},
+            {"Vaal Flesh Offering", "Active Skill Gem"},
+            {"Vaal Glacial Hammer", "Active Skill Gem"},
+            {"Vaal Grace", "Active Skill Gem"},
+            {"Vaal Ground Slam", "Active Skill Gem"},
+            {"Vaal Haste", "Active Skill Gem"},
+            {"Vaal Heavy Strike", "Active Skill Gem"},
+            {"Vaal Ice Nova", "Active Skill Gem"},
+            {"Vaal Immortal Call", "Active Skill Gem"},
+            {"Vaal Impurity of Lightning", "Active Skill Gem"},
+            {"Vaal Lightning Strike", "Active Skill Gem"},
+            {"Vaal Lightning Trap", "Active Skill Gem"},
+            {"Vaal Lightning Warp", "Active Skill Gem"},
+            {"Vaal Molten Shell", "Active Skill Gem"},
+            {"Vaal Breach", "Active Skill Gem"},
+            {"Vaal Power Siphon", "Active Skill Gem"},
+            {"Vaal Rain of Arrows", "Active Skill Gem"},
+            {"Vaal Reave", "Active Skill Gem"},
+            {"Vaal Righteous Fire", "Active Skill Gem"},
+            {"Vaal Spark", "Active Skill Gem"},
+            {"Vaal Storm Call", "Active Skill Gem"},
+            {"Vaal Summon Skeletons", "Active Skill Gem"},
+            {"Vaal Sweep", "Active Skill Gem"},
+            {"Vaal Spectral Throw", "Active Skill Gem"},
+            {"Vengeance", "Active Skill Gem"},
+            {"Venom Gyre", "Active Skill Gem"},
+            {"Vigilant Strike", "Active Skill Gem"},
+            {"Viper Strike", "Active Skill Gem"},
+            {"Vitality", "Active Skill Gem"},
+            {"Void Sphere", "Active Skill Gem"},
+            {"Volatile Dead", "Active Skill Gem"},
+            {"Vortex Mine", "Active Skill Gem"},
+            {"Despair", "Active Skill Gem"},
+            {"Wand Teleport", "Active Skill Gem"},
+            {"War Banner", "Active Skill Gem"},
+            {"Warlord's Mark", "Active Skill Gem"},
+            {"Whirling Blades", "Active Skill Gem"},
+            {"Wild Strike", "Active Skill Gem"},
+            {"Winter Orb", "Active Skill Gem"},
+            {"Wintertide Brand", "Active Skill Gem"},
+            {"Wither", "Active Skill Gem"},
+            {"Withering Step", "Active Skill Gem"},
+            {"Wrath", "Active Skill Gem"},
+            {"Added Chaos Damage Support", "Support Skill Gem"},
+            {"Awakened Added Chaos Damage Support", "Support Skill Gem"},
+            {"Added Cold Damage Support", "Support Skill Gem"},
+            {"Awakened Added Cold Damage Support", "Support Skill Gem"},
+            {"Added Fire Damage Support", "Support Skill Gem"},
+            {"Awakened Added Fire Damage Support", "Support Skill Gem"},
+            {"Added Lightning Damage Support", "Support Skill Gem"},
+            {"Awakened Added Lightning Damage Support", "Support Skill Gem"},
+            {"Additional Accuracy Support", "Support Skill Gem"},
+            {"Empower Support", "Support Skill Gem"},
+            {"Enhance Support", "Support Skill Gem"},
+            {"Enlighten Support", "Support Skill Gem"},
+            {"Awakened Ancestral Call Support", "Support Skill Gem"},
+            {"Arcane Surge Support", "Support Skill Gem"},
+            {"Archmage Support", "Support Skill Gem"},
+            {"Arrow Nova Support", "Support Skill Gem"},
+            {"Awakened Arrow Nova Support", "Support Skill Gem"},
+            {"Barrage Support", "Support Skill Gem"},
+            {"Blasphemy Support", "Support Skill Gem"},
+            {"Awakened Blasphemy Support", "Support Skill Gem"},
+            {"Blind Support", "Support Skill Gem"},
+            {"Block Chance Reduction Support", "Support Skill Gem"},
+            {"Arrogance Support", "Support Skill Gem"},
+            {"Bloodlust Support", "Support Skill Gem"},
+            {"Bloodthirst Support", "Support Skill Gem"},
+            {"Bonechill Support", "Support Skill Gem"},
+            {"Brutality Support", "Support Skill Gem"},
+            {"Awakened Brutality Support", "Support Skill Gem"},
+            {"Awakened Burning Damage Support", "Support Skill Gem"},
+            {"WIP Support", "Support Skill Gem"},
+            {"Cast On Critical Strike Support", "Support Skill Gem"},
+            {"Awakened Cast On Critical Strike Support", "Support Skill Gem"},
+            {"Cast when Damage Taken Support", "Support Skill Gem"},
+            {"Cast on Death Support", "Support Skill Gem"},
+            {"Cast on Melee Kill Support", "Support Skill Gem"},
+            {"Cast when Stunned Support", "Support Skill Gem"},
+            {"Cast while Channelling Support", "Support Skill Gem"},
+            {"Awakened Cast While Channelling Support", "Support Skill Gem"},
+            {"Chain Support", "Support Skill Gem"},
+            {"Awakened Chain Support", "Support Skill Gem"},
+            {"Chance to Bleed Support", "Support Skill Gem"},
+            {"Chance to Flee Support", "Support Skill Gem"},
+            {"Combustion Support", "Support Skill Gem"},
+            {"Withering Touch Support", "Support Skill Gem"},
+            {"Charged Mines Support", "Support Skill Gem"},
+            {"Close Combat Support", "Support Skill Gem"},
+            {"Cluster Traps Support", "Support Skill Gem"},
+            {"Cold Penetration Support", "Support Skill Gem"},
+            {"Awakened Cold Penetration Support", "Support Skill Gem"},
+            {"Cold to Fire Support", "Support Skill Gem"},
+            {"Concentrated Effect Support", "Support Skill Gem"},
+            {"Controlled Destruction Support", "Support Skill Gem"},
+            {"Awakened Controlled Destruction Support", "Support Skill Gem"},
+            {"Cruelty Support", "Support Skill Gem"},
+            {"Culling Strike Support", "Support Skill Gem"},
+            {"Hextouch Support", "Support Skill Gem"},
+            {"Awakened Hextouch Support", "Support Skill Gem"},
+            {"Hypothermia Support", "Support Skill Gem"},
+            {"Deadly Ailments Support", "Support Skill Gem"},
+            {"Awakened Deadly Ailments Support", "Support Skill Gem"},
+            {"Predator Support", "Support Skill Gem"},
+            {"Decay Support", "Support Skill Gem"},
+            {"Efficacy Support", "Support Skill Gem"},
+            {"Elemental Focus Support", "Support Skill Gem"},
+            {"Awakened Elemental Focus Support", "Support Skill Gem"},
+            {"Elemental Penetration Support", "Support Skill Gem"},
+            {"Elemental Proliferation Support", "Support Skill Gem"},
+            {"Endurance Charge on Melee Stun Support", "Support Skill Gem"},
+            {"Energy Leech Support", "Support Skill Gem"},
+            {"Faster Attacks Support", "Support Skill Gem"},
+            {"Faster Casting Support", "Support Skill Gem"},
+            {"Faster Projectiles Support", "Support Skill Gem"},
+            {"Feeding Frenzy Support", "Support Skill Gem"},
+            {"Fire Penetration Support", "Support Skill Gem"},
+            {"Awakened Fire Penetration Support", "Support Skill Gem"},
+            {"Fist of War Support", "Support Skill Gem"},
+            {"Fork Support", "Support Skill Gem"},
+            {"Awakened Fork Support", "Support Skill Gem"},
+            {"Fortify Support", "Support Skill Gem"},
+            {"Ice Bite Support", "Support Skill Gem"},
+            {"Charged Traps Support", "Support Skill Gem"},
+            {"Generosity Support", "Support Skill Gem"},
+            {"Awakened Generosity Support", "Support Skill Gem"},
+            {"Greater Multiple Projectiles Support", "Support Skill Gem"},
+            {"Awakened Greater Multiple Projectiles Support", "Support Skill Gem"},
+            {"Greater Volley Support", "Support Skill Gem"},
+            {"Unleash Support", "Support Skill Gem"},
+            {"[UNUSED] Handcast Rapid Fire", "Support Skill Gem"},
+            {"Intensify Support", "Support Skill Gem"},
+            {"High-Impact Mine Support", "Support Skill Gem"},
+            {"Ignite Proliferation Support", "Support Skill Gem"},
+            {"Immolate Support", "Support Skill Gem"},
+            {"Impale Support", "Support Skill Gem"},
+            {"Impending Doom Support", "Support Skill Gem"},
+            {"Increased Area of Effect Support", "Support Skill Gem"},
+            {"Awakened Increased Area of Effect Support", "Support Skill Gem"},
+            {"Burning Damage Support", "Support Skill Gem"},
+            {"Increased Critical Damage Support", "Support Skill Gem"},
+            {"Increased Critical Strikes Support", "Support Skill Gem"},
+            {"Increased Duration Support", "Support Skill Gem"},
+            {"Infernal Legion Support", "Support Skill Gem"},
+            {"Iron Grip Support", "Support Skill Gem"},
+            {"Iron Will Support", "Support Skill Gem"},
+            {"Item Quantity Support", "Support Skill Gem"},
+            {"Item Rarity Support", "Support Skill Gem"},
+            {"Knockback Support", "Support Skill Gem"},
+            {"Lesser Multiple Projectiles Support", "Support Skill Gem"},
+            {"Lesser Poison Support", "Support Skill Gem"},
+            {"Life Gain on Hit Support", "Support Skill Gem"},
+            {"Life Leech Support", "Support Skill Gem"},
+            {"Lifetap Support", "Support Skill Gem"},
+            {"Lightning Penetration Support", "Support Skill Gem"},
+            {"Awakened Lightning Penetration Support", "Support Skill Gem"},
+            {"Maim Support", "Support Skill Gem"},
+            {"Mana Leech Support", "Support Skill Gem"},
+            {"Meat Shield Support", "Support Skill Gem"},
+            {"Damage on Full Life Support", "Support Skill Gem"},
+            {"Melee Physical Damage Support", "Support Skill Gem"},
+            {"Awakened Melee Physical Damage Support", "Support Skill Gem"},
+            {"Melee Splash Support", "Support Skill Gem"},
+            {"Awakened Melee Splash Support", "Support Skill Gem"},
+            {"Minefield Support", "Support Skill Gem"},
+            {"Minion Damage Support", "Support Skill Gem"},
+            {"Awakened Minion Damage Support", "Support Skill Gem"},
+            {"Minion Life Support", "Support Skill Gem"},
+            {"Minion Speed Support", "Support Skill Gem"},
+            {"Mirage Archer Support", "Support Skill Gem"},
+            {"Multiple Totems Support", "Support Skill Gem"},
+            {"Multiple Traps Support", "Support Skill Gem"},
+            {"Spell Echo Support", "Support Skill Gem"},
+            {"Multistrike Support", "Support Skill Gem"},
+            {"Awakened Multistrike Support", "Support Skill Gem"},
+            {"Nightblade Support", "Support Skill Gem"},
+            {"Onslaught Support", "Support Skill Gem"},
+            {"Innervate Support", "Support Skill Gem"},
+            {"Volley Support", "Support Skill Gem"},
+            {"Vicious Projectiles Support", "Support Skill Gem"},
+            {"Physical to Lightning Support", "Support Skill Gem"},
+            {"Pierce Support", "Support Skill Gem"},
+            {"Pinpoint Support", "Support Skill Gem"},
+            {"Point Blank Support", "Support Skill Gem"},
+            {"Poison Support", "Support Skill Gem"},
+            {"Power Charge On Critical Support", "Support Skill Gem"},
+            {"Pulverise Support", "Support Skill Gem"},
+            {"Rage Support", "Support Skill Gem"},
+            {"Ballista Totem Support", "Support Skill Gem"},
+            {"Swift Affliction Support", "Support Skill Gem"},
+            {"Less Duration Support", "Support Skill Gem"},
+            {"Inspiration Support", "Support Skill Gem"},
+            {"Blastchain Mine Support", "Support Skill Gem"},
+            {"Return Projectiles Support", "Support Skill Gem"},
+            {"Ruthless Support", "Support Skill Gem"},
+            {"Second Wind Support", "Support Skill Gem"},
+            {"Shockwave Support", "Support Skill Gem"},
+            {"Slower Projectiles Support", "Support Skill Gem"},
+            {"Spell Cascade Support", "Support Skill Gem"},
+            {"Awakened Spell Cascade Support", "Support Skill Gem"},
+            {"Awakened Spell Echo Support", "Support Skill Gem"},
+            {"Spell Totem Support", "Support Skill Gem"},
+            {"Ancestral Call Support", "Support Skill Gem"},
+            {"Split Projectiles Support", "Support Skill Gem"},
+            {"Infused Channelling Support", "Support Skill Gem"},
+            {"Stun Support", "Support Skill Gem"},
+            {"Elemental Army Support", "Support Skill Gem"},
+            {"Summon Phantasm Support", "Support Skill Gem"},
+            {"Awakened Swift Affliction Support", "Support Skill Gem"},
+            {"Swift Assembly Support", "Support Skill Gem"},
+            {"Swiftbrand Support", "Support Skill Gem"},
+            {"Lesser Reduced Mana Cost Support", "Support Skill Gem"},
+            {"Trap Support", "Support Skill Gem"},
+            {"Trap and Mine Damage Support", "Support Skill Gem"},
+            {"Advanced Traps Support", "Support Skill Gem"},
+            {"Trinity Support", "Support Skill Gem"},
+            {"Unbound Ailments Support", "Support Skill Gem"},
+            {"Awakened Unbound Ailments Support", "Support Skill Gem"},
+            {"Awakened Unleash Support", "Support Skill Gem"},
+            {"Urgent Orders Support", "Support Skill Gem"},
+            {"Vaal Soul Harvesting", "Support Skill Gem"},
+            {"Awakened Vicious Projectiles Support", "Support Skill Gem"},
+            {"Vile Toxins Support", "Support Skill Gem"},
+            {"Void Manipulation Support", "Support Skill Gem"},
+            {"Awakened Void Manipulation Support", "Support Skill Gem"},
+            {"Elemental Damage with Attacks Support", "Support Skill Gem"},
+            {"Awakened Elemental Damage with Attacks Support", "Support Skill Gem"},
+            {"Rogue's Marker", "StackableCurrency"},
+            {"Hypnotic Eye Jewel", "AbyssJewel"},
+            {"Murderous Eye Jewel", "AbyssJewel"},
+            {"Searching Eye Jewel", "AbyssJewel"},
+            {"Ghastly Eye Jewel", "AbyssJewel"},
+            {"Viridian Jewel", "Jewel"},
+            {"Cobalt Jewel", "Jewel"},
+            {"Large Cluster Jewel", "Jewel"},
+            {"Medium Cluster Jewel", "Jewel"},
+            {"Small Cluster Jewel", "Jewel"},
+            {"Prismatic Jewel", "Jewel"},
+            {"Crimson Jewel", "Jewel"},
+            {"Timeless Jewel", "Jewel"},
+            {"Simulacrum", "MapFragment"},
+            {"Chayula's Breachstone", "MapFragment"},
+            {"Chayula's Charged Breachstone", "MapFragment"},
+            {"Chayula's Enriched Breachstone", "MapFragment"},
+            {"Chayula's Pure Breachstone", "MapFragment"},
+            {"Tul's Breachstone", "MapFragment"},
+            {"Tul's Charged Breachstone", "MapFragment"},
+            {"Tul's Enriched Breachstone", "MapFragment"},
+            {"Tul's Pure Breachstone", "MapFragment"},
+            {"Xoph's Breachstone", "MapFragment"},
+            {"Xoph's Charged Breachstone", "MapFragment"},
+            {"Xoph's Enriched Breachstone", "MapFragment"},
+            {"Xoph's Pure Breachstone", "MapFragment"},
+            {"Esh's Breachstone", "MapFragment"},
+            {"Esh's Charged Breachstone", "MapFragment"},
+            {"Esh's Enriched Breachstone", "MapFragment"},
+            {"Esh's Pure Breachstone", "MapFragment"},
+            {"Uul-Netol's Breachstone", "MapFragment"},
+            {"Uul-Netol's Charged Breachstone", "MapFragment"},
+            {"Uul-Netol's Enriched Breachstone", "MapFragment"},
+            {"Uul-Netol's Pure Breachstone", "MapFragment"},
+            {"Fragment of Enslavement", "MapFragment"},
+            {"Fragment of Eradication", "MapFragment"},
+            {"Fragment of Constriction", "MapFragment"},
+            {"Fragment of Purification", "MapFragment"},
+            {"Key to Decay", "MapFragment"},
+            {"Divine Vessel", "MapFragment"},
+            {"Sacred Blossom", "MapFragment"},
+            {"Timeless Eternal Emblem", "MapFragment"},
+            {"Timeless Karui Emblem", "MapFragment"},
+            {"Timeless Maraketh Emblem", "MapFragment"},
+            {"Timeless Templar Emblem", "MapFragment"},
+            {"Timeless Vaal Emblem", "MapFragment"},
+            {"The Maven's Writ", "MapFragment"},
+            {"Crescent Splinter", "StackableCurrency"},
+            {"Offering to the Goddess", "MapFragment"},
+            {"Dedication to the Goddess", "MapFragment"},
+            {"Gift to the Goddess", "MapFragment"},
+            {"Tribute to the Goddess", "MapFragment"},
+            {"Eber's Key", "MapFragment"},
+            {"Yriel's Key", "MapFragment"},
+            {"Inya's Key", "MapFragment"},
+            {"Volkuur's Key", "MapFragment"},
+            {"Pale Court Set", "MapFragment"},
+            {"Fragment of the Phoenix", "MapFragment"},
+            {"Fragment of the Minotaur", "MapFragment"},
+            {"Fragment of the Chimera", "MapFragment"},
+            {"Fragment of the Hydra", "MapFragment"},
+            {"Key to the Crucible", "MapFragment"},
+            {"Fragment of Terror", "MapFragment"},
+            {"Fragment of Emptiness", "MapFragment"},
+            {"Fragment of Shape", "MapFragment"},
+            {"Fragment of Knowledge", "MapFragment"},
+            {"Maddening Object", "MapFragment"},
+            {"Sacrifice at Midnight", "MapFragment"},
+            {"Sacrifice at Dawn", "MapFragment"},
+            {"Sacrifice at Noon", "MapFragment"},
+            {"Sacrifice at Dusk", "MapFragment"},
+            {"Mortal Rage", "MapFragment"},
+            {"Mortal Hope", "MapFragment"},
+            {"Mortal Ignorance", "MapFragment"},
+            {"Mortal Grief", "MapFragment"},
+            {"Sacrifice Set", "MapFragment"},
+            {"Mortal Set", "MapFragment"},
+            {"Blood-filled Vessel", "MapFragment"},
+            {"Necropolis Map", "Map"},
+            {"Plateau Map", "Map"},
+            {"Bazaar Map", "Map"},
+            {"Volcano Map", "Map"},
+            {"Chateau Map", "Map"},
+            {"Lava Chamber Map", "Map"},
+            {"Precinct Map", "Map"},
+            {"Academy Map", "Map"},
+            {"Fungal Hollow Map", "Map"},
+            {"Shipyard Map", "Map"},
+            {"Overgrown Ruin Map", "Map"},
+            {"Castle Ruins Map", "Map"},
+            {"Arsenal Map", "Map"},
+            {"Wasteland Map", "Map"},
+            {"Courtyard Map", "Map"},
+            {"Excavation Map", "Map"},
+            {"Waterways Map", "Map"},
+            {"Palace Map", "Map"},
+            {"Shrine Map", "Map"},
+            {"Maze Map", "Map"},
+            {"Vaal Temple Map", "Map"},
+            {"Plaza Map", "Map"},
+            {"Core Map", "Map"},
+            {"Caldera Map", "Map"},
+            {"Colosseum Map", "Map"},
+            {"Cursed Crypt Map", "Map"},
+            {"Dungeon Map", "Map"},
+            {"Grotto Map", "Map"},
+            {"Dunes Map", "Map"},
+            {"Pit Map", "Map"},
+            {"Tropical Island Map", "Map"},
+            {"Desert Map", "Map"},
+            {"Toxic Sewer Map", "Map"},
+            {"Channel Map", "Map"},
+            {"Thicket Map", "Map"},
+            {"Atoll Map", "Map"},
+            {"Cemetery Map", "Map"},
+            {"Arcade Map", "Map"},
+            {"Wharf Map", "Map"},
+            {"Ghetto Map", "Map"},
+            {"Spider Lair Map", "Map"},
+            {"Vaal Pyramid Map", "Map"},
+            {"Reef Map", "Map"},
+            {"Geode Map", "Map"},
+            {"Mud Geyser Map", "Map"},
+            {"Museum Map", "Map"},
+            {"Arena Map", "Map"},
+            {"Overgrown Shrine Map", "Map"},
+            {"Crystal Ore Map", "Map"},
+            {"Shore Map", "Map"},
+            {"Spider Forest Map", "Map"},
+            {"Promenade Map", "Map"},
+            {"Phantasmagoria Map", "Map"},
+            {"Underground River Map", "Map"},
+            {"Pier Map", "Map"},
+            {"Bog Map", "Map"},
+            {"Graveyard Map", "Map"},
+            {"Coves Map", "Map"},
+            {"Villa Map", "Map"},
+            {"Temple Map", "Map"},
+            {"Arachnid Nest Map", "Map"},
+            {"Strand Map", "Map"},
+            {"Ashen Wood Map", "Map"},
+            {"Colonnade Map", "Map"},
+            {"Bone Crypt Map", "Map"},
+            {"Primordial Blocks Map", "Map"},
+            {"Waste Pool Map", "Map"},
+            {"Flooded Mine Map", "Map"},
+            {"Jungle Valley Map", "Map"},
+            {"Gardens Map", "Map"},
+            {"Cells Map", "Map"},
+            {"Canyon Map", "Map"},
+            {"Dark Forest Map", "Map"},
+            {"Peninsula Map", "Map"},
+            {"Orchard Map", "Map"},
+            {"Underground Sea Map", "Map"},
+            {"Arid Lake Map", "Map"},
+            {"Glacier Map", "Map"},
+            {"Residence Map", "Map"},
+            {"Malformation Map", "Map"},
+            {"Shaped Academy Map", "Map"},
+            {"Acid Caverns Map", "Map"},
+            {"Shaped Acid Caverns Map", "Map"},
+            {"Shaped Arachnid Nest Map", "Map"},
+            {"Arachnid Tomb Map", "Map"},
+            {"Shaped Arachnid Tomb Map", "Map"},
+            {"Shaped Arcade Map", "Map"},
+            {"Shaped Arena Map", "Map"},
+            {"Shaped Arid Lake Map", "Map"},
+            {"Armoury Map", "Map"},
+            {"Shaped Armoury Map", "Map"},
+            {"Shaped Arsenal Map", "Map"},
+            {"Shaped Atoll Map", "Map"},
+            {"Barrows Map", "Map"},
+            {"Shaped Barrows Map", "Map"},
+            {"Beach Map", "Map"},
+            {"Shaped Beach Map", "Map"},
+            {"Lighthouse Map", "Map"},
+            {"Shaped Bog Map", "Map"},
+            {"Burial Chambers Map", "Map"},
+            {"Shaped Burial Chambers Map", "Map"},
+            {"Shaped Canyon Map", "Map"},
+            {"Shaped Castle Ruins Map", "Map"},
+            {"Shaped Bone Crypt Map", "Map"},
+            {"Shaped Flooded Mine Map", "Map"},
+            {"Shaped Cells Map", "Map"},
+            {"Shaped Cemetery Map", "Map"},
+            {"Shaped Channel Map", "Map"},
+            {"Pit of the Chimera Map", "Map"},
+            {"Shaped Colonnade Map", "Map"},
+            {"Shaped Courtyard Map", "Map"},
+            {"Shaped Coves Map", "Map"},
+            {"Shaped Cursed Crypt Map", "Map"},
+            {"Shaped Crystal Ore Map", "Map"},
+            {"Shaped Desert Map", "Map"},
+            {"Shaped Peninsula Map", "Map"},
+            {"Shaped Ashen Wood Map", "Map"},
+            {"Shaped Dunes Map", "Map"},
+            {"Shaped Dungeon Map", "Map"},
+            {"Estuary Map", "Map"},
+            {"Factory Map", "Map"},
+            {"Shaped Factory Map", "Map"},
+            {"Shaped Ghetto Map", "Map"},
+            {"Shaped Graveyard Map", "Map"},
+            {"Shaped Grotto Map", "Map"},
+            {"Harbinger Map", "Map"},
+            {"Terrace Map", "Map"},
+            {"Lair of the Hydra Map", "Map"},
+            {"Ivory Temple Map", "Map"},
+            {"Shaped Jungle Valley Map", "Map"},
+            {"Lair Map", "Map"},
+            {"Shaped Malformation Map", "Map"},
+            {"Marshes Map", "Map"},
+            {"Shaped Marshes Map", "Map"},
+            {"Mesa Map", "Map"},
+            {"Shaped Mesa Map", "Map"},
+            {"Mineral Pools Map", "Map"},
+            {"Maze of the Minotaur Map", "Map"},
+            {"Shaped Mud Geyser Map", "Map"},
+            {"Shaped Museum Map", "Map"},
+            {"Desert Spring Map", "Map"},
+            {"Shaped Desert Spring Map", "Map"},
+            {"Shaped Orchard Map", "Map"},
+            {"Shaped Overgrown Shrine Map", "Map"},
+            {"Shaped Phantasmagoria Map", "Map"},
+            {"Forge of the Phoenix Map", "Map"},
+            {"Shaped Pier Map", "Map"},
+            {"Shaped Pit Map", "Map"},
+            {"Primordial Pool Map", "Map"},
+            {"Shaped Primordial Pool Map", "Map"},
+            {"Shaped Promenade Map", "Map"},
+            {"Shaped Geode Map", "Map"},
+            {"Port Map", "Map"},
+            {"Shaped Port Map", "Map"},
+            {"Racecourse Map", "Map"},
+            {"Shaped Racecourse Map", "Map"},
+            {"Ramparts Map", "Map"},
+            {"Shaped Ramparts Map", "Map"},
+            {"Shaped Reef Map", "Map"},
+            {"Scriptorium Map", "Map"},
+            {"Shaped Toxic Sewer Map", "Map"},
+            {"The Shaper's Realm", "Map"},
+            {"Shaped Shore Map", "Map"},
+            {"Shaped Spider Forest Map", "Map"},
+            {"Shaped Spider Lair Map", "Map"},
+            {"Shaped Strand Map", "Map"},
+            {"Leyline Map", "Map"},
+            {"Shaped Temple Map", "Map"},
+            {"Shaped Gardens Map", "Map"},
+            {"Shaped Thicket Map", "Map"},
+            {"Tower Map", "Map"},
+            {"Shaped Tower Map", "Map"},
+            {"Shaped Tropical Island Map", "Map"},
+            {"Shaped Underground River Map", "Map"},
+            {"Ancient City Map", "Map"},
+            {"Shaped Ancient City Map", "Map"},
+            {"Shaped Vaal Pyramid Map", "Map"},
+            {"Vault Map", "Map"},
+            {"Shaped Villa Map", "Map"},
+            {"Shaped Waste Pool Map", "Map"},
+            {"Shaped Wharf Map", "Map"},
+            {"Alleyways Map", "Map"},
+            {"Basilica Map", "Map"},
+            {"Belfry Map", "Map"},
+            {"Bramble Valley Map", "Map"},
+            {"Cage Map", "Map"},
+            {"Carcass Map", "Map"},
+            {"City Square Map", "Map"},
+            {"Cold River Map", "Map"},
+            {"Conservatory Map", "Map"},
+            {"Coral Ruins Map", "Map"},
+            {"Courthouse Map", "Map"},
+            {"Crimson Temple Map", "Map"},
+            {"Crimson Township Map", "Map"},
+            {"Defiled Cathedral Map", "Map"},
+            {"Dig Map", "Map"},
+            {"Dry Sea Map", "Map"},
+            {"Fields Map", "Map"},
+            {"Forbidden Woods Map", "Map"},
+            {"Forking River Map", "Map"},
+            {"Foundry Map", "Map"},
+            {"Frozen Cabins Map", "Map"},
+            {"Grave Trough Map", "Map"},
+            {"Haunted Mansion Map", "Map"},
+            {"Iceberg Map", "Map"},
+            {"Infested Valley Map", "Map"},
+            {"Laboratory Map", "Map"},
+            {"Lava Lake Map", "Map"},
+            {"Lookout Map", "Map"},
+            {"Mausoleum Map", "Map"},
+            {"Moon Temple Map", "Map"},
+            {"Park Map", "Map"},
+            {"Pen Map", "Map"},
+            {"Relic Chambers Map", "Map"},
+            {"Sepulchre Map", "Map"},
+            {"Siege Map", "Map"},
+            {"Silo Map", "Map"},
+            {"Stagnation Map", "Map"},
+            {"Sulphur Vents Map", "Map"},
+            {"Summit Map", "Map"},
+            {"Sunken City Map", "Map"},
+            {"Crater Map", "Map"},
+            {"Cured Quiver", "Quiver"},
+            {"Fire Arrow Quiver", "Quiver"},
+            {"Broadhead Arrow Quiver", "Quiver"},
+            {"Penetrating Arrow Quiver", "Quiver"},
+            {"Spike-Point Arrow Quiver", "Quiver"},
+            {"Ornate Quiver", "Quiver"},
+            {"Rugged Quiver", "Quiver"},
+            {"Conductive Quiver", "Quiver"},
+            {"Heavy Quiver", "Quiver"},
+            {"Light Quiver", "Quiver"},
+            {"Serrated Arrow Quiver", "Quiver"},
+            {"Two-Point Arrow Quiver", "Quiver"},
+            {"Sharktooth Arrow Quiver", "Quiver"},
+            {"Blunt Arrow Quiver", "Quiver"},
+            {"Artillery Quiver", "Quiver"},
+            {"Breach Ring", "Ring"},
+            {"Iron Ring", "Ring"},
+            {"Amethyst Ring", "Ring"},
+            {"Diamond Ring", "Ring"},
+            {"Two-Stone Ring", "Ring"},
+            {"Unset Ring", "Ring"},
+            {"Coral Ring", "Ring"},
+            {"Paua Ring", "Ring"},
+            {"Gold Ring", "Ring"},
+            {"Topaz Ring", "Ring"},
+            {"Sapphire Ring", "Ring"},
+            {"Ruby Ring", "Ring"},
+            {"Prismatic Ring", "Ring"},
+            {"Moonstone Ring", "Ring"},
+            {"Steel Ring", "Ring"},
+            {"Opal Ring", "Ring"},
+            {"Vermillion Ring", "Ring"},
+            {"Cerulean Ring", "Ring"},
+            {"Iolite Ring", "Ring"},
+            {"Golden Hoop", "Ring"},
+            {"Cogwork Ring", "Ring"},
+            {"Geodesic Ring", "Ring"},
+            {"Jet Ring", "Ring"},
+            {"Rusted Abyss Scarab", "MapFragment"},
+            {"Polished Abyss Scarab", "MapFragment"},
+            {"Gilded Abyss Scarab", "MapFragment"},
+            {"Winged Abyss Scarab", "MapFragment"},
+            {"Rusted Bestiary Scarab", "MapFragment"},
+            {"Polished Bestiary Scarab", "MapFragment"},
+            {"Gilded Bestiary Scarab", "MapFragment"},
+            {"Winged Bestiary Scarab", "MapFragment"},
+            {"Farric Lure", "MapFragment"},
+            {"Saqawine Lure", "MapFragment"},
+            {"Fenumal Lure", "MapFragment"},
+            {"Craicic Lure", "MapFragment"},
+            {"Rusted Blight Scarab", "MapFragment"},
+            {"Polished Blight Scarab", "MapFragment"},
+            {"Gilded Blight Scarab", "MapFragment"},
+            {"Winged Blight Scarab", "MapFragment"},
+            {"Rusted Breach Scarab", "MapFragment"},
+            {"Polished Breach Scarab", "MapFragment"},
+            {"Gilded Breach Scarab", "MapFragment"},
+            {"Winged Breach Scarab", "MapFragment"},
+            {"Rusted Divination Scarab", "MapFragment"},
+            {"Polished Divination Scarab", "MapFragment"},
+            {"Gilded Divination Scarab", "MapFragment"},
+            {"Winged Divination Scarab", "MapFragment"},
+            {"Rusted Elder Scarab", "MapFragment"},
+            {"Polished Elder Scarab", "MapFragment"},
+            {"Gilded Elder Scarab", "MapFragment"},
+            {"Winged Elder Scarab", "MapFragment"},
+            {"Rusted Harbinger Scarab", "MapFragment"},
+            {"Polished Harbinger Scarab", "MapFragment"},
+            {"Gilded Harbinger Scarab", "MapFragment"},
+            {"Winged Harbinger Scarab", "MapFragment"},
+            {"Rusted Legion Scarab", "MapFragment"},
+            {"Polished Legion Scarab", "MapFragment"},
+            {"Gilded Legion Scarab", "MapFragment"},
+            {"Winged Legion Scarab", "MapFragment"},
+            {"Rusted Cartography Scarab", "MapFragment"},
+            {"Polished Cartography Scarab", "MapFragment"},
+            {"Gilded Cartography Scarab", "MapFragment"},
+            {"Winged Cartography Scarab", "MapFragment"},
+            {"Rusted Metamorph Scarab", "MapFragment"},
+            {"Polished Metamorph Scarab", "MapFragment"},
+            {"Gilded Metamorph Scarab", "MapFragment"},
+            {"Winged Metamorph Scarab", "MapFragment"},
+            {"Rusted Perandus Scarab", "MapFragment"},
+            {"Polished Perandus Scarab", "MapFragment"},
+            {"Gilded Perandus Scarab", "MapFragment"},
+            {"Winged Perandus Scarab", "MapFragment"},
+            {"Rusted Shaper Scarab", "MapFragment"},
+            {"Polished Shaper Scarab", "MapFragment"},
+            {"Gilded Shaper Scarab", "MapFragment"},
+            {"Winged Shaper Scarab", "MapFragment"},
+            {"Rusted Ambush Scarab", "MapFragment"},
+            {"Polished Ambush Scarab", "MapFragment"},
+            {"Gilded Ambush Scarab", "MapFragment"},
+            {"Winged Ambush Scarab", "MapFragment"},
+            {"Rusted Sulphite Scarab", "MapFragment"},
+            {"Polished Sulphite Scarab", "MapFragment"},
+            {"Gilded Sulphite Scarab", "MapFragment"},
+            {"Winged Sulphite Scarab", "MapFragment"},
+            {"Rusted Torment Scarab", "MapFragment"},
+            {"Polished Torment Scarab", "MapFragment"},
+            {"Gilded Torment Scarab", "MapFragment"},
+            {"Winged Torment Scarab", "MapFragment"},
+            {"Rusted Reliquary Scarab", "MapFragment"},
+            {"Polished Reliquary Scarab", "MapFragment"},
+            {"Gilded Reliquary Scarab", "MapFragment"},
+            {"Winged Reliquary Scarab", "MapFragment"},
+            {"Nailed Fist", "Claw"},
+            {"Gouger", "Claw"},
+            {"Tiger's Paw", "Claw"},
+            {"Gut Ripper", "Claw"},
+            {"Prehistoric Claw", "Claw"},
+            {"Noble Claw", "Claw"},
+            {"Eagle Claw", "Claw"},
+            {"Great White Claw", "Claw"},
+            {"Throat Stabber", "Claw"},
+            {"Hellion's Paw", "Claw"},
+            {"Eye Gouger", "Claw"},
+            {"Sharktooth Claw", "Claw"},
+            {"Vaal Claw", "Claw"},
+            {"Imperial Claw", "Claw"},
+            {"Terror Claw", "Claw"},
+            {"Awl", "Claw"},
+            {"Cat's Paw", "Claw"},
+            {"Blinder", "Claw"},
+            {"Timeworn Claw", "Claw"},
+            {"Sparkling Claw", "Claw"},
+            {"Fright Claw", "Claw"},
+            {"Thresher Claw", "Claw"},
+            {"Shadow Fangs", "Claw"},
+            {"Malign Fangs", "Claw"},
+            {"Void Fangs", "Claw"},
+            {"Double Claw", "Claw"},
+            {"Twin Claw", "Claw"},
+            {"Gemini Claw", "Claw"},
+            {"Glass Shank", "Dagger"},
+            {"Butcher Knife", "Rune Dagger"},
+            {"Poignard", "Dagger"},
+            {"Boot Blade", "Rune Dagger"},
+            {"Golden Kris", "Rune Dagger"},
+            {"Royal Skean", "Rune Dagger"},
+            {"Fiend Dagger", "Rune Dagger"},
+            {"Gutting Knife", "Dagger"},
+            {"Slaughter Knife", "Rune Dagger"},
+            {"Ambusher", "Dagger"},
+            {"Ezomyte Dagger", "Rune Dagger"},
+            {"Skinning Knife", "Dagger"},
+            {"Platinum Kris", "Rune Dagger"},
+            {"Imperial Skean", "Rune Dagger"},
+            {"Demon Dagger", "Rune Dagger"},
+            {"Carving Knife", "Rune Dagger"},
+            {"Stiletto", "Dagger"},
+            {"Boot Knife", "Rune Dagger"},
+            {"Copper Kris", "Rune Dagger"},
+            {"Skean", "Rune Dagger"},
+            {"Imp Dagger", "Rune Dagger"},
+            {"Flaying Knife", "Dagger"},
+            {"Hollowpoint Dagger", "Dagger"},
+            {"Pressurised Dagger", "Dagger"},
+            {"Pneumatic Dagger", "Dagger"},
+            {"Prong Dagger", "Dagger"},
+            {"Trisula", "Dagger"},
+            {"Sai", "Dagger"},
+            {"Ethereal Blade", "Dagger"},
+            {"Flickerflame Blade", "Rune Dagger"},
+            {"Flashfire Blade", "Rune Dagger"},
+            {"Infernal Blade", "Rune Dagger"},
+            {"Rusted Hatchet", "One Hand Axe"},
+            {"Tomahawk", "One Hand Axe"},
+            {"Wrist Chopper", "One Hand Axe"},
+            {"War Axe", "One Hand Axe"},
+            {"Chest Splitter", "One Hand Axe"},
+            {"Ceremonial Axe", "One Hand Axe"},
+            {"Wraith Axe", "One Hand Axe"},
+            {"Karui Axe", "One Hand Axe"},
+            {"Siege Axe", "One Hand Axe"},
+            {"Reaver Axe", "One Hand Axe"},
+            {"Butcher Axe", "One Hand Axe"},
+            {"Jade Hatchet", "One Hand Axe"},
+            {"Vaal Hatchet", "One Hand Axe"},
+            {"Royal Axe", "One Hand Axe"},
+            {"Infernal Axe", "One Hand Axe"},
+            {"Boarding Axe", "One Hand Axe"},
+            {"Cleaver", "One Hand Axe"},
+            {"Broad Axe", "One Hand Axe"},
+            {"Arming Axe", "One Hand Axe"},
+            {"Decorative Axe", "One Hand Axe"},
+            {"Spectral Axe", "One Hand Axe"},
+            {"Jasper Axe", "One Hand Axe"},
+            {"Maltreatment Axe", "One Hand Axe"},
+            {"Disapprobation Axe", "One Hand Axe"},
+            {"Psychotic Axe", "One Hand Axe"},
+            {"Etched Hatchet", "One Hand Axe"},
+            {"Engraved Hatchet", "One Hand Axe"},
+            {"Runic Hatchet", "One Hand Axe"},
+            {"Driftwood Club", "One Hand Mace"},
+            {"Barbed Club", "One Hand Mace"},
+            {"Rock Breaker", "One Hand Mace"},
+            {"Battle Hammer", "One Hand Mace"},
+            {"Flanged Mace", "One Hand Mace"},
+            {"Ornate Mace", "One Hand Mace"},
+            {"Phantom Mace", "One Hand Mace"},
+            {"Ancestral Club", "One Hand Mace"},
+            {"Tenderizer", "One Hand Mace"},
+            {"Gavel", "One Hand Mace"},
+            {"Legion Hammer", "One Hand Mace"},
+            {"Tribal Club", "One Hand Mace"},
+            {"Pernach", "One Hand Mace"},
+            {"Auric Mace", "One Hand Mace"},
+            {"Nightmare Mace", "One Hand Mace"},
+            {"Spiked Club", "One Hand Mace"},
+            {"Stone Hammer", "One Hand Mace"},
+            {"War Hammer", "One Hand Mace"},
+            {"Bladed Mace", "One Hand Mace"},
+            {"Ceremonial Mace", "One Hand Mace"},
+            {"Dream Mace", "One Hand Mace"},
+            {"Petrified Club", "One Hand Mace"},
+            {"Flare Mace", "One Hand Mace"},
+            {"Crack Mace", "One Hand Mace"},
+            {"Boom Mace", "One Hand Mace"},
+            {"Wyrm Mace", "One Hand Mace"},
+            {"Dragon Mace", "One Hand Mace"},
+            {"Behemoth Mace", "One Hand Mace"},
+            {"Driftwood Sceptre", "Sceptre"},
+            {"Sekhem", "Sceptre"},
+            {"Crystal Sceptre", "Sceptre"},
+            {"Lead Sceptre", "Sceptre"},
+            {"Blood Sceptre", "Sceptre"},
+            {"Royal Sceptre", "Sceptre"},
+            {"Abyssal Sceptre", "Sceptre"},
+            {"Karui Sceptre", "Sceptre"},
+            {"Tyrant's Sekhem", "Sceptre"},
+            {"Opal Sceptre", "Sceptre"},
+            {"Platinum Sceptre", "Sceptre"},
+            {"Darkwood Sceptre", "Sceptre"},
+            {"Vaal Sceptre", "Sceptre"},
+            {"Carnal Sceptre", "Sceptre"},
+            {"Void Sceptre", "Sceptre"},
+            {"Bronze Sceptre", "Sceptre"},
+            {"Quartz Sceptre", "Sceptre"},
+            {"Iron Sceptre", "Sceptre"},
+            {"Ochre Sceptre", "Sceptre"},
+            {"Ritual Sceptre", "Sceptre"},
+            {"Shadow Sceptre", "Sceptre"},
+            {"Grinning Fetish", "Sceptre"},
+            {"Oscillating Sceptre", "Sceptre"},
+            {"Stabilising Sceptre", "Sceptre"},
+            {"Alternating Sceptre", "Sceptre"},
+            {"Horned Sceptre", "Sceptre"},
+            {"Stag Sceptre", "Sceptre"},
+            {"Sambar Sceptre", "Sceptre"},
+            {"Rusted Sword", "One Hand Sword"},
+            {"Cutlass", "One Hand Sword"},
+            {"Baselard", "One Hand Sword"},
+            {"Battle Sword", "One Hand Sword"},
+            {"Elder Sword", "One Hand Sword"},
+            {"Graceful Sword", "One Hand Sword"},
+            {"Twilight Blade", "One Hand Sword"},
+            {"Gemstone Sword", "One Hand Sword"},
+            {"Corsair Sword", "One Hand Sword"},
+            {"Gladius", "One Hand Sword"},
+            {"Legion Sword", "One Hand Sword"},
+            {"Copper Sword", "One Hand Sword"},
+            {"Vaal Blade", "One Hand Sword"},
+            {"Eternal Sword", "One Hand Sword"},
+            {"Midnight Blade", "One Hand Sword"},
+            {"Sabre", "One Hand Sword"},
+            {"Broad Sword", "One Hand Sword"},
+            {"War Sword", "One Hand Sword"},
+            {"Ancient Sword", "One Hand Sword"},
+            {"Elegant Sword", "One Hand Sword"},
+            {"Dusk Blade", "One Hand Sword"},
+            {"Variscite Blade", "One Hand Sword"},
+            {"Fickle Spiritblade", "One Hand Sword"},
+            {"Capricious Spiritblade", "One Hand Sword"},
+            {"Anarchic Spiritblade", "One Hand Sword"},
+            {"Hook Sword", "One Hand Sword"},
+            {"Grappler", "One Hand Sword"},
+            {"Tiger Hook", "One Hand Sword"},
+            {"Rusted Spike", "Thrusting One Hand Sword"},
+            {"Burnished Foil", "Thrusting One Hand Sword"},
+            {"Estoc", "Thrusting One Hand Sword"},
+            {"Serrated Foil", "Thrusting One Hand Sword"},
+            {"Primeval Rapier", "Thrusting One Hand Sword"},
+            {"Fancy Foil", "Thrusting One Hand Sword"},
+            {"Apex Rapier", "Thrusting One Hand Sword"},
+            {"Dragonbone Rapier", "Thrusting One Hand Sword"},
+            {"Tempered Foil", "Thrusting One Hand Sword"},
+            {"Pecoraro", "Thrusting One Hand Sword"},
+            {"Spiraled Foil", "Thrusting One Hand Sword"},
+            {"Whalebone Rapier", "Thrusting One Hand Sword"},
+            {"Vaal Rapier", "Thrusting One Hand Sword"},
+            {"Jewelled Foil", "Thrusting One Hand Sword"},
+            {"Harpy Rapier", "Thrusting One Hand Sword"},
+            {"Battered Foil", "Thrusting One Hand Sword"},
+            {"Basket Rapier", "Thrusting One Hand Sword"},
+            {"Jagged Foil", "Thrusting One Hand Sword"},
+            {"Antique Rapier", "Thrusting One Hand Sword"},
+            {"Elegant Foil", "Thrusting One Hand Sword"},
+            {"Thorn Rapier", "Thrusting One Hand Sword"},
+            {"Wyrmbone Rapier", "Thrusting One Hand Sword"},
+            {"Smallsword", "Thrusting One Hand Sword"},
+            {"Courtesan Sword", "Thrusting One Hand Sword"},
+            {"Dragoon Sword", "Thrusting One Hand Sword"},
+            {"Driftwood Wand", "Wand"},
+            {"Serpent Wand", "Wand"},
+            {"Omen Wand", "Wand"},
+            {"Demon's Horn", "Wand"},
+            {"Imbued Wand", "Wand"},
+            {"Opal Wand", "Wand"},
+            {"Tornado Wand", "Wand"},
+            {"Prophecy Wand", "Wand"},
+            {"Goat's Horn", "Wand"},
+            {"Carved Wand", "Wand"},
+            {"Quartz Wand", "Wand"},
+            {"Spiraled Wand", "Wand"},
+            {"Sage Wand", "Wand"},
+            {"Faun's Horn", "Wand"},
+            {"Engraved Wand", "Wand"},
+            {"Crystal Wand", "Wand"},
+            {"Convoking Wand", "Wand"},
+            {"Assembler Wand", "Wand"},
+            {"Congregator Wand", "Wand"},
+            {"Accumulator Wand", "Wand"},
+            {"Pagan Wand", "Wand"},
+            {"Heathen Wand", "Wand"},
+            {"Profane Wand", "Wand"},
+            {"Fishing Rod", "FishingRod"},
+            {"Crude Bow", "Bow"},
+            {"Decurve Bow", "Bow"},
+            {"Compound Bow", "Bow"},
+            {"Sniper Bow", "Bow"},
+            {"Ivory Bow", "Bow"},
+            {"Highborn Bow", "Bow"},
+            {"Decimation Bow", "Bow"},
+            {"Thicket Bow", "Bow"},
+            {"Citadel Bow", "Bow"},
+            {"Ranger Bow", "Bow"},
+            {"Assassin Bow", "Bow"},
+            {"Short Bow", "Bow"},
+            {"Spine Bow", "Bow"},
+            {"Imperial Bow", "Bow"},
+            {"Harbinger Bow", "Bow"},
+            {"Long Bow", "Bow"},
+            {"Composite Bow", "Bow"},
+            {"Recurve Bow", "Bow"},
+            {"Bone Bow", "Bow"},
+            {"Royal Bow", "Bow"},
+            {"Death Bow", "Bow"},
+            {"Grove Bow", "Bow"},
+            {"Hedron Bow", "Bow"},
+            {"Foundry Bow", "Bow"},
+            {"Solarine Bow", "Bow"},
+            {"Reflex Bow", "Bow"},
+            {"Steelwood Bow", "Bow"},
+            {"Maraketh Bow", "Bow"},
+            {"Gnarled Branch", "Staff"},
+            {"Military Staff", "Warstaff"},
+            {"Serpentine Staff", "Warstaff"},
+            {"Highborn Staff", "Staff"},
+            {"Foul Staff", "Warstaff"},
+            {"Primordial Staff", "Staff"},
+            {"Lathi", "Staff"},
+            {"Ezomyte Staff", "Warstaff"},
+            {"Maelström Staff", "Warstaff"},
+            {"Imperial Staff", "Staff"},
+            {"Judgement Staff", "Warstaff"},
+            {"Primitive Staff", "Staff"},
+            {"Long Staff", "Staff"},
+            {"Iron Staff", "Warstaff"},
+            {"Coiled Staff", "Warstaff"},
+            {"Royal Staff", "Staff"},
+            {"Vile Staff", "Warstaff"},
+            {"Woodful Staff", "Staff"},
+            {"Quarterstaff", "Staff"},
+            {"Transformer Staff", "Staff"},
+            {"Reciprocation Staff", "Staff"},
+            {"Battery Staff", "Staff"},
+            {"Crescent Staff", "Staff"},
+            {"Moon Staff", "Staff"},
+            {"Eclipse Staff", "Staff"},
+            {"Capacity Rod", "Warstaff"},
+            {"Potentiality Rod", "Warstaff"},
+            {"Eventuality Rod", "Warstaff"},
+            {"Stone Axe", "Two Hand Axe"},
+            {"Headsman Axe", "Two Hand Axe"},
+            {"Labrys", "Two Hand Axe"},
+            {"Noble Axe", "Two Hand Axe"},
+            {"Abyssal Axe", "Two Hand Axe"},
+            {"Karui Chopper", "Two Hand Axe"},
+            {"Sundering Axe", "Two Hand Axe"},
+            {"Ezomyte Axe", "Two Hand Axe"},
+            {"Vaal Axe", "Two Hand Axe"},
+            {"Despot Axe", "Two Hand Axe"},
+            {"Void Axe", "Two Hand Axe"},
+            {"Jade Chopper", "Two Hand Axe"},
+            {"Woodsplitter", "Two Hand Axe"},
+            {"Poleaxe", "Two Hand Axe"},
+            {"Double Axe", "Two Hand Axe"},
+            {"Gilded Axe", "Two Hand Axe"},
+            {"Shadow Axe", "Two Hand Axe"},
+            {"Jasper Chopper", "Two Hand Axe"},
+            {"Timber Axe", "Two Hand Axe"},
+            {"Prime Cleaver", "Two Hand Axe"},
+            {"Honed Cleaver", "Two Hand Axe"},
+            {"Apex Cleaver", "Two Hand Axe"},
+            {"Dagger Axe", "Two Hand Axe"},
+            {"Talon Axe", "Two Hand Axe"},
+            {"Fleshripper", "Two Hand Axe"},
+            {"Driftwood Maul", "Two Hand Mace"},
+            {"Steelhead", "Two Hand Mace"},
+            {"Spiny Maul", "Two Hand Mace"},
+            {"Plated Maul", "Two Hand Mace"},
+            {"Dread Maul", "Two Hand Mace"},
+            {"Karui Maul", "Two Hand Mace"},
+            {"Colossus Mallet", "Two Hand Mace"},
+            {"Piledriver", "Two Hand Mace"},
+            {"Meatgrinder", "Two Hand Mace"},
+            {"Imperial Maul", "Two Hand Mace"},
+            {"Terror Maul", "Two Hand Mace"},
+            {"Tribal Maul", "Two Hand Mace"},
+            {"Mallet", "Two Hand Mace"},
+            {"Sledgehammer", "Two Hand Mace"},
+            {"Jagged Maul", "Two Hand Mace"},
+            {"Brass Maul", "Two Hand Mace"},
+            {"Fright Maul", "Two Hand Mace"},
+            {"Totemic Maul", "Two Hand Mace"},
+            {"Great Mallet", "Two Hand Mace"},
+            {"Blunt Force Condenser", "Two Hand Mace"},
+            {"Crushing Force Magnifier", "Two Hand Mace"},
+            {"Impact Force Propagator", "Two Hand Mace"},
+            {"Morning Star", "Two Hand Mace"},
+            {"Solar Maul", "Two Hand Mace"},
+            {"Coronal Maul", "Two Hand Mace"},
+            {"Corroded Blade", "Two Hand Sword"},
+            {"Highland Blade", "Two Hand Sword"},
+            {"Engraved Greatsword", "Two Hand Sword"},
+            {"Tiger Sword", "Two Hand Sword"},
+            {"Wraith Sword", "Two Hand Sword"},
+            {"Headman's Sword", "Two Hand Sword"},
+            {"Reaver Sword", "Two Hand Sword"},
+            {"Ezomyte Blade", "Two Hand Sword"},
+            {"Vaal Greatsword", "Two Hand Sword"},
+            {"Lion Sword", "Two Hand Sword"},
+            {"Infernal Sword", "Two Hand Sword"},
+            {"Longsword", "Two Hand Sword"},
+            {"Bastard Sword", "Two Hand Sword"},
+            {"Two-Handed Sword", "Two Hand Sword"},
+            {"Etched Greatsword", "Two Hand Sword"},
+            {"Ornate Sword", "Two Hand Sword"},
+            {"Spectral Sword", "Two Hand Sword"},
+            {"Butcher Sword", "Two Hand Sword"},
+            {"Footman Sword", "Two Hand Sword"},
+            {"Keyblade", "Two Hand Sword"},
+            {"Rebuking Blade", "Two Hand Sword"},
+            {"Blasting Blade", "Two Hand Sword"},
+            {"Banishing Blade", "Two Hand Sword"},
+            {"Curved Blade", "Two Hand Sword"},
+            {"Lithe Blade", "Two Hand Sword"},
+            {"Exquisite Blade", "Two Hand Sword"},
+            {"Charan's Sword", "One Hand Sword"},
+            {"Golden Blade", "One Hand Sword"}
+        };
+
+
+        
+
+
+        // Try this later in another refactor, maybe quicker? Maybe not if we use reflection
+        //public static string AbyssJewel => "Abyss Jewel";
+
+        public static Dictionary<string, string> ItemClassToRawItemClassName = new()
+        {
+            {"AbyssJewel", "Abyss Jewel"},
+            {"Active Skill Gem", "Active Skill Gems"},
+            {"Amulet", "Amulets"},
+            {"AtlasRegionUpgradeItem", "Atlas Region Upgrade Item"},
+            {"Belt", "Belts"},
+            {"Body Armour", "Body Armours"},
+            {"Boots", "Boots"},
+            {"Bow", "Bows"},
+            {"Claw", "Claws"},
+            {"Currency", "Currency"},
+            {"Dagger", "Daggers"},
+            {"DelveSocketableCurrency", "Delve Socketable Currency"},
+            {"DelveStackableSocketableCurrency", "Delve Stackable Socketable Currency"},
+            {"DivinationCard", "Divination Card"},
+            {"FishingRod", "Fishing Rods"},
+            {"Gloves", "Gloves"},
+            {"HeistBlueprint", "Blueprint"},
+            {"HeistContract", "Contract"},
+            {"HeistEquipmentReward", "Heist Brooch"},
+            {"HeistEquipmentTool", "Heist Tool"},
+            {"HeistEquipmentUtility", "Heist Cloak"},
+            {"HeistEquipmentWeapon", "Heist Gear"},
+            {"HeistObjective", "Heist Target"},
+            {"Helmet", "Helmets"},
+            {"HiddenItem", "Hidden Item"},
+            {"HideoutDoodad", "Hideout Doodads"},
+            {"HybridFlask", "Hybrid Flasks"},
+            {"Incubator", "Incubator"},
+            {"IncursionItem", "Incursion Item"},
+            {"Jewel", "Jewel"},
+            {"LabyrinthItem", "Labyrinth Item"},
+            {"LabyrinthMapItem", "Labyrinth Map Item"},
+            {"LabyrinthTrinket", "Labyrinth Trinket"},
+            {"LargeRelic", "Large Relics"},
+            {"Leaguestone", "Leaguestones"},
+            {"LifeFlask", "Life Flasks"},
+            {"ManaFlask", "Mana Flasks"},
+            {"Map", "Maps"},
+            {"MapFragment", "Map Fragments"},
+            {"MediumRelic", "Medium Relics"},
+            {"MetamorphosisDNA", "Metamorph Sample"},
+            {"Microtransaction", "Microtransactions"},
+            {"MiscMapItem", "Misc Map Items"},
+            {"One Hand Axe", "One Hand Axes"},
+            {"One Hand Mace", "One Hand Maces"},
+            {"One Hand Sword", "One Hand Swords"},
+            {"PantheonSoul", "Pantheon Soul"},
+            {"QuestItem", "Quest Items"},
+            {"Quiver", "Quivers"},
+            {"Ring", "Rings"},
+            {"Rune Dagger", "Rune Daggers"},
+            {"Sceptre", "Sceptres"},
+            {"Shield", "Shields"},
+            {"SmallRelic", "Small Relics"},
+            {"StackableCurrency", "Stackable Currency"},
+            {"Staff", "Staves"},
+            {"Support Skill Gem", "Support Skill Gems"},
+            {"Thrusting One Hand Sword", "Thrusting One Hand Swords"},
+            {"Trinket", "Trinkets"},
+            {"Two Hand Axe", "Two Hand Axes"},
+            {"Two Hand Mace", "Two Hand Maces"},
+            {"Two Hand Sword", "Two Hand Swords"},
+            {"UNUSED2", ""},
+            {"UNUSED3", ""},
+            {"UNUSED4", ""},
+            {"Unarmed", ""},
+            {"UniqueFragment", "Piece"},
+            {"UniqueShard", "Shard"},
+            {"UniqueShardBase", "Shard Heart"},
+            {"UtilityFlask", "Utility Flasks"},
+            {"UtilityFlaskCritical", "Critical Utility Flasks"},
+            {"Wand", "Wands"},
+            {"Warstaff", "Warstaves"}
+        };
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+        public static HashSet<string> ArmourCategories = new()
+        {
+            "Garb",
+            "Jacket",
+            "Cage",
+            "Dragonscale",
+            "Leather",
+            "Flame",
+            "Burgonet",
+            "Greaves",
+            "Buckler",
+            "Sallet",
+            "Raiment",
+            "Mask",
+            "Shield",
+            "Crown",
+            "Gloves",
+            "Mitts",
+            "Doublet",
+            "Plate",
+            "Tricorne",
+            "Circlet",
+            "Gauntlets",
+            "Shoes",
+            "Vestment",
+            "Cap",
+            "Quiver",
+            "Brigandine",
+            "Helmet",
+            "Helm",
+            "Hauberk",
+            "Bascinet",
+            "Slippers",
+            "Chainmail",
+            "Vest",
+            "Regalia",
+            "Lamellar",
+            "Silks",
+            "Ringmail",
+            "Coat",
+            "Tunic",
+            "Armour",
+            "Mantle",
+            "Boots",
+            "Hood",
+            "Pelt",
+            "Robe",
+            "Hat",
+            "Wyrmscale"
+        };
+       
+        
+
+
+
+        public static HashSet<string> AccessoryCategories = new HashSet<string>() 
+        { 
+            "Vise",
+            "Ring",
+            "Amulet",
+            "Talisman",
+            "Belt",
+            "Rustic Sash"
+        };
+
+
+
+
+        public static HashSet<string> WeaponCategories = new HashSet<string>() 
+        {   
+            "Sword",
+            "Tomahawk",
+            "Branch",
+            "Knife",
+            "Cutlass",
+            "Greatsword",
+            "Awl",
+            "Paw",
+            "Breaker",
+            "Sledgehammer",
+            "Axe",
+            "Chopper",
+            "Hammer",
+            "Dagger",
+            "Estoc",
+            "Ripper",
+            "Maul",
+            "Lathi",
+            "Blade",
+            "Labrys",
+            "Sekhem",
+            "Club",
+            "Foil",
+            "Wand",
+            "Fetish",
+            "Skean",
+            "Gladius",
+            "Sceptre",
+            "Woodsplitter",
+            "Steelhead",
+            "Gavel",
+            "Poleaxe",
+            "Sabre",
+            "Staff",
+            "Ambusher",
+            "Cleaver",
+            "Blinder",
+            "Horn",
+            "Stiletto",
+            "Stabber",
+            "Meatgrinder",
+            "Fist",
+            "Mallet",
+            "Hatchet",
+            "Bow",
+            "Mace",
+            "Rapier",
+            "Claw",
+            "Kris"
+        };
     }
 }
